% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mixingmetric.R
\name{add_mixingmetric}
\alias{add_mixingmetric}
\title{Add mixing metric}
\usage{
add_mixingmetric(
  obj,
  reduction,
  new_col = paste0("MixingMetric_", reduction),
  grouping.var = "source",
  k = 5,
  max.k = 300,
  dims = NULL,
  invert = TRUE,
  normalise = FALSE,
  ...
)
}
\arguments{
\item{reduction}{Which dimensionally reduced space to use}

\item{grouping.var}{Grouping variable for dataset}

\item{k}{Neighbor number to examine per group}

\item{max.k}{Maximum size of local neighborhood to compute}

\item{dims}{Dimensions to use}

\item{invert}{Invert the mixing metric such that higher values 
indicate better mixing.}

\item{normalise}{Normalise the mixing metric to the range [0,1].}

\item{...}{
  Arguments passed on to \code{\link[Seurat:MixingMetric]{Seurat::MixingMetric}}
  \describe{
    \item{\code{object}}{Seurat object}
    \item{\code{eps}}{Error bound on the neighbor finding algorithm (from RANN)}
    \item{\code{verbose}}{Displays progress bar}
  }}
}
\description{
Compute the mixing metric for a given reduction and add it to the metadata 
of a \link{Seurat} object. See \link[Seurat]{MixingMetric} for more details.
}
