% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_zstat_magma.R
\name{adjust_zstat_magma}
\alias{adjust_zstat_magma}
\title{Adjust GWAS Z-statistic: MAGMA}
\usage{
adjust_zstat_magma(
  magma,
  drop_MHC = TRUE,
  method = "bonferroni",
  model = NULL,
  formula = ZSTAT ~ NSNPS + logNSNPS + NPARAM + logNPARAM + GENELEN + logGENELEN,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{magma}{MAGMA \emph{.genes.out} data.table
produced by \link[MAGMA.Celltyping]{map.snps.to.genes}.}

\item{drop_MHC}{Drop genes from the MHC 
(Major Histocompatibility Complex) region.}

\item{method}{correction method, a \code{\link{character}} string.
    Can be abbreviated.}

\item{model}{Statistical model to use. 
Defaults to \link[stats]{lm}.}

\item{formula}{Formula to use in \code{model}.}

\item{...}{Additional arguments passed to \code{model}.}
}
\description{
Adjust MAGMA Z-statistic from \emph{.genes.out} files
}
\details{
Used when you want to directly analyse the gene-level 
Z-scores corrected for gene length,  etc.
}
