% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_zstat_txdb.R
\name{adjust_zstat_txdb}
\alias{adjust_zstat_txdb}
\title{Adjust GWAS Z-statistic: txdb}
\usage{
adjust_zstat_txdb(
  dat,
  drop_MHC = TRUE,
  method = "bonferroni",
  model = NULL,
  formula = ZSTAT ~ NSNPS + logNSNPS + GENELEN + logGENELEN,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{\code{data.table}
produced by \link[phenomix]{map_snps2genes_txdb}.}

\item{drop_MHC}{Drop genes from the MHC
(Major Histocompatibility Complex) region.}

\item{method}{correction method, a \code{\link{character}} string.
    Can be abbreviated.}

\item{model}{Statistical model to use.
Defaults to \link[stats]{lm}.}

\item{formula}{Formula to use in \code{model}.}

\item{...}{Additional arguments passed to \code{model}.}
}
\value{
\code{dat} with the new column "ADJ_ZSTAT".
}
\description{
Adjust the GWAS Z-statistic for confounders like gene length.
}
\details{
Used when you want to directly analyse the gene-level
Z-scores corrected for gene length,  etc.
}
