% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sumstats.R
\name{aggregate_sumstats}
\alias{aggregate_sumstats}
\title{Aggregate summary statistics}
\source{
\href{https://stackoverflow.com/questions/59059743/apply-different-functions-to-different-columns-programmatically-in-data-table-r}{data.table aggregation solution}
}
\usage{
aggregate_sumstats(
  merged_hits,
  drop_na = FALSE,
  agg_var = "SYMBOL",
  extra_cols = c("ABC.Score", "CellType"),
  sort_rows = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{merged_hits}{\link[data.table]{data.table}
with the columns "BETA", "P","SE", and \code{agg_var}.}

\item{drop_na}{Drop rows with \code{NA} in any column.}

\item{agg_var}{Variable to aggregate \code{merged_hits} by.}

\item{extra_cols}{Extra columns to include.}

\item{sort_rows}{Sort rows alphanumerically.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} with
mean and standard deviation of "BETA", "P", and "SE", as well as
number of rows and number of unique SNPs per grouping level.
}
\description{
Efficiently aggregate GWAS/QTL summary statistics
according to some grouping column (e.g."GENEID").
}
\details{
Requires the columns
}
\keyword{internal}
