% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cores.r
\name{assign_cores}
\alias{assign_cores}
\title{Assign cores}
\usage{
assign_cores(workers = 0.9, progressbar = TRUE, verbose = TRUE)
}
\arguments{
\item{workers}{
    \code{integer(1)} Number of workers. Defaults to the maximum of 1 or
    the number of cores determined by \code{detectCores} minus 2 unless
    environment variables \code{R_PARALLELLY_AVAILABLECORES_FALLBACK} or
    \code{BIOCPARALLEL_WORKER_NUMBER} are set otherwise.
  }

\item{progressbar}{
    \code{logical(1)} Enable progress bar (based on
    plyr:::progress_text). Enabling the progress bar changes the
    \emph{default} value of \code{tasks} to \code{.Machine$integer.max}, so
    that progress is reported for each element of \code{X}.
  }
}
\description{
Automatically infer the number of available cores and
 allocate the majority of them to parallel processing.
 Reserves a small proportion of cores to avoid slowing or crashing.
}
\keyword{internal}
