% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chum}
\alias{chum}
\title{Count data collected by Washington Department of Fish and Wildlife on
chum salmon from the Skagit River (Washington state). Each row of the
dataframe contains an observation ("number") on a given date ("date").
The year ("year") and calendar day ("doy") are also included.}
\format{
A data frame.
}
\usage{
chum
}
\description{
Count data collected by Washington Department of Fish and Wildlife on
chum salmon from the Skagit River (Washington state). Each row of the
dataframe contains an observation ("number") on a given date ("date").
The year ("year") and calendar day ("doy") are also included.
}
\keyword{internal}
