% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cor.R
\name{compute_cor}
\alias{compute_cor}
\title{Compute all pairwise trait correlations}
\usage{
compute_cor(
  obj,
  assay = NULL,
  slot = NULL,
  reduction = NULL,
  transpose = FALSE,
  method = "pearson",
  return_obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Either a \pkg{Seurat} object or a feature*trait matrix.}

\item{assay}{If \code{obj} is a \pkg{Seurat} object, which assay to extract
for correlations.}

\item{slot}{If \code{obj} is a \pkg{Seurat} object, which slot to extract
for correlations.}

\item{reduction}{If not \code{NULL}, extracts the embedding from the
specified reduction (e.g. "pca") and uses that to
compute trait-trait similarities instead.}

\item{transpose}{Whether to transpose the matrix first.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{return_obj}{Whether to return the \pkg{Seurat} object with a new
\code{obj@graphs} slot, or to simply return the sparse correlation matrix.}

\item{verbose}{Print messages.}
}
\value{
A \pkg{Seurat} object with a new \code{obj@graphs} slot,
or the sparse correlation matrix.
}
\description{
Computes pairwise correlations for all traits
 using either \link[WGCNA]{cor} (faster) or
\link[stats]{cor} (slower).
}
