% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Create data file for fitting time varying run timing distributions with TMB}
\usage{
create_data(
  data,
  min_number = 0,
  variable = "number",
  time = "year",
  date = "doy",
  asymmetric_model = TRUE,
  sigma_trend = TRUE,
  mu_trend = TRUE,
  family = "gaussian"
)
}
\arguments{
\item{data}{A data frame}

\item{min_number}{A minimum threshold to use, defaults to 0}

\item{variable}{A character string of the name of the variable in 'data' that contains the response (e.g. counts)}

\item{time}{A character string of the name of the variable in 'data' that contains the time variable (e.g. year)}

\item{date}{A character string of the name of the variable in 'data' that contains the response (e.g. day of year). The actual
column should contain a numeric response -- for example, the result from using lubridate::yday(x)}

\item{asymmetric_model}{Boolean, whether or not to let model be asymmetric (e.g. run timing before peak has a
different shape than run timing after peak)}

\item{sigma_trend}{Whether to fit a log-linear trend in standard deviations. Defaults to TRUE (when FALSE, an intercept and random deviations are estimated)}

\item{mu_trend}{Whether to fit a linear trend in means. Defaults to TRUE (when FALSE, an intercept and random deviations are estimated)}

\item{family}{Response for observation model, options are "gaussian", "poisson", "negbin"}
}
\description{
Does minimal processing of data to use as argument to fitting function
}
\examples{
data(fishdist)
datalist = create_data(fishdist, min_number = 0, variable = "number", time = "year",
date = "doy", asymmetric_model = TRUE, family = "gaussian")
}
