% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DimReducObject.R
\name{create_DimReducObject}
\alias{create_DimReducObject}
\title{Create a \code{DimReducObject} from DeGAs}
\usage{
create_DimReducObject(reconstruction, assay, key)
}
\arguments{
\item{reconstruction}{Output of \code{reconstruct_matrix}
with embeddings and loadings.}

\item{assay}{Assay to use.}

\item{key}{Key to use (name of embedding).}
}
\description{
Create a \pkg{Seurat} \code{DimReducObject}
from DeGAs embeddings and loadings.
}
\keyword{internal}
