% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DT.R
\name{create_DT}
\alias{create_DT}
\title{Create an interactive data table with download buttons}
\usage{
create_DT(DF, caption = "", scrollY = 400)
}
\arguments{
\item{DF}{Data table.}

\item{caption}{Table caption.}

\item{scrollY}{Max height in pixels before scrolling.}
}
\description{
Create an interactive data table with download buttons
}
\keyword{internal}
