% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ddgnorm}
\alias{ddgnorm}
\title{Calculate the density of a double general normal distribution}
\usage{
ddgnorm(x, mu, alpha1, alpha2, beta1, beta2, sigma1, sigma2)
}
\arguments{
\item{x}{The value to evaluate the density at}

\item{mu}{The mean of the distribution}

\item{alpha1}{The alpha parameter of the left half of the distribution}

\item{alpha2}{The alpha parameter of the right half of the distribution}

\item{beta1}{The beta parameter of the left half of the distribution}

\item{beta2}{The beta parameter of the right half of the distribution}

\item{sigma1}{The sigma of the left half of the distribution}

\item{sigma2}{The sigma of the right half of the distribution}
}
\description{
Calculate the density of a double general normal distribution
}
