% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ddnorm}
\alias{ddnorm}
\title{Calculate the density of a double normal distribution}
\usage{
ddnorm(x, mu, sigma1, sigma2)
}
\arguments{
\item{x}{The value to evaluate the density at}

\item{mu}{The mean of the distribution}

\item{sigma1}{The std deviation of the left half of the distribution}

\item{sigma2}{The std deviation of the right half of the distribution}
}
\description{
Calculate the density of a double normal distribution
}
