% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ddt}
\alias{ddt}
\title{Calculate the density of a double Student-t distribution}
\usage{
ddt(x, mu, sigma1, sigma2, tdf_1, tdf_2)
}
\arguments{
\item{x}{The value to evaluate the density at}

\item{mu}{The mean of the distribution}

\item{sigma1}{The std deviation of the left half of the distribution}

\item{sigma2}{The std deviation of the right half of the distribution}

\item{tdf_1}{The df deviation of the left half of the distribution}

\item{tdf_2}{The df deviation of the right half of the distribution}
}
\description{
Calculate the density of a double Student-t distribution
}
