% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEGAS_contributionGene}
\alias{DEGAS_contributionGene}
\title{DeGAs gene loadings}
\format{
An object of class \code{DimReduc} with 2138 rows and 100 columns.
}
\source{
\href{https://github.com/rivas-lab/public-resources/tree/master/uk_biobank/DeGAs}{DeGAs GitHub}
}
\usage{
DEGAS_contributionGene
}
\description{
DeGAs gene loadings
}
\examples{
\dontrun{
DEGAS_contributionGene <- readRDS("/Desktop/phenome_decomposition/raw_data/DEGAS/contributionGene.rds")
usethis::use_data(DEGAS_contributionGene, overwrite = TRUE)
}
}
\keyword{datasets}
