% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_granges.R
\name{dt_to_granges}
\alias{dt_to_granges}
\title{Convert data.table to GRanges object}
\usage{
dt_to_granges(
  dat,
  chrom_col = "CHR",
  start_col = "POS",
  end_col = start_col,
  style = "NCBI",
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data.}

\item{chrom_col}{Chromosome column name.}

\item{start_col}{Genomic start position column name.}

\item{end_col}{Genomic end position column name.}

\item{style}{GRanges style (e.g. "NCBI, "UCSC") 
set by \link[GenomeInfoDb]{seqlevelsStyle}.}

\item{verbose}{Print messages.}
}
\description{
Convert \link[data.table]{data.table}/\link[base]{data.frame} to a 
\link[GenomicRanges]{GRanges} object.
}
