% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_all}
\alias{extract_all}
\title{Output processing function to be called by user}
\usage{
extract_all(fit)
}
\arguments{
\item{fit}{A fitted object returned from fit()}
}
\description{
This function extracts the means, sigmas, thetas,
lower (25\%) and upper (75\%) quartiles, and respective sds
}
