% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_annual}
\alias{extract_annual}
\title{Output processing function to be called by user}
\usage{
extract_annual(fit, log = TRUE)
}
\arguments{
\item{fit}{A fitted object returned from fit()}

\item{log}{Whether to return estimates in log space, defaults to TRUE}
}
\description{
This function extracts the annual totals
}
