% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cor.R
\name{extract_cor}
\alias{extract_cor}
\title{Extract correlation matrix}
\usage{
extract_cor(
  obj,
  assay = NULL,
  slot = NULL,
  reduction = NULL,
  graph_name = NULL,
  method = "pearson",
  verbose = TRUE
)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{Seurat}}
}}

\item{assay}{Assay to used to compute correlation graph 
if one does not already exist.}

\item{slot}{Name of component object to return}

\item{reduction}{Which reduction to extract from.
Only used if \code{obj} is of class \code{Seurat}.}

\item{graph_name}{Name of the graph to extract.}

\item{method}{Pairwise correlation method.}

\item{verbose}{Print messages.}
}
\value{
Trait-trait correlation matrix.
}
\description{
Extract correlation matrix stored as a graph.
}
