% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_embeddings.R
\name{extract_embeddings}
\alias{extract_embeddings}
\title{Extract embeddings}
\usage{
extract_embeddings(obj, reduction = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{Seurat}}
}}

\item{reduction}{Which reduction to extract from. 
Only used if \code{obj} is of class \code{Seurat}.}

\item{verbose}{Print messages.}
}
\description{
Extract embeddings from various object types.
}
