% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_loadings.R
\name{extract_loadings}
\alias{extract_loadings}
\title{Extract loadings}
\usage{
extract_loadings(obj, reduction = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{prcomp}}
\item{\code{list} : }{Generated by \code{run_[tools]} decomposition function.}
}}

\item{reduction}{Which reduction to extract from.
Defaults to first available reduction.
Only used if \code{obj} is of class \code{Seurat}.}

\item{verbose}{Print messages.}
}
\description{
Extract loadings from various object types.
}
