% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_lower}
\alias{extract_lower}
\title{Output processing function to be called by user}
\usage{
extract_lower(fit)
}
\arguments{
\item{fit}{A fitted object returned from fit()}
}
\description{
This function extracts the lower quartiles (25\%) and respective sds
}
