% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_means}
\alias{extract_means}
\title{Output processing function to be called by user}
\usage{
extract_means(fit)
}
\arguments{
\item{fit}{A fitted object returned from fit()}
}
\description{
This function extracts the parameter means and respective sds
}
