% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sigma}
\alias{extract_sigma}
\title{Output processing function to be called by user}
\usage{
extract_sigma(fit)
}
\arguments{
\item{fit}{A fitted object returned from fit()}
}
\description{
This function extracts the parameter sigma and respective sds
}
