% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_theta}
\alias{extract_theta}
\title{Output processing function to be called by user}
\usage{
extract_theta(fit)
}
\arguments{
\item{fit}{A fitted object returned from fit()}
}
\description{
This function extracts the parameter theta and respective sds
}
