% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_upper}
\alias{extract_upper}
\title{Output processing function to be called by user}
\usage{
extract_upper(fit)
}
\arguments{
\item{fit}{A fitted object returned from fit()}
}
\description{
This function extracts the upper quartiles (25\%) and respective sds
}
