% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.phenomix}
\alias{fitted.phenomix}
\title{Get fitted values from model object, copying glmmTMB 'fast' implementation}
\usage{
\method{fitted}{phenomix}(object, ...)
}
\arguments{
\item{...}{Extra parameters}

\item{fit}{The fitted phenomix model}
}
\description{
Get fitted values from model object, copying glmmTMB 'fast' implementation
}
