% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fixef.phenomix}
\alias{fixef.phenomix}
\title{Get fixed effects parameters from model object, copying glmmTMB 'fast' implementation}
\usage{
\method{fixef}{phenomix}(object, ...)
}
\arguments{
\item{object}{The fitted phenomix model}

\item{...}{Additional arguments}
}
\description{
Get fixed effects parameters from model object, copying glmmTMB 'fast' implementation
}
