% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata.R
\name{generate_metadata}
\alias{generate_metadata}
\title{Generate metadata}
\usage{
generate_metadata(
  munged_files,
  infer_builds = TRUE,
  drop_na_N = TRUE,
  trait_groups = NULL,
  force_new = FALSE,
  save_dir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{munged_files}{List of paths to summary statostics files
that have been munged with \link[MungeSumstats]{import_sumstats}.}

\item{infer_builds}{Infer genome builds of each file.}

\item{drop_na_N}{Drop rows without "N" (total sample size).}

\item{trait_groups}{[Optional] A vector of strings.
Creates a new column "trait_groups"
based on a series of substring searches.}

\item{force_new}{Create a new metadata file even when one
already exists in \code{save_dir}.}

\item{save_dir}{Directory to save metadata to.}

\item{verbose}{Print messages.}
}
\value{
data.frame of OpenGWAS metadata.
}
\description{
Generate metadata for \href{https://gwas.mrcieu.ac.uk/}{OpenGWAS} datasets.
}
