% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BlueLake2018_FrontalCortexOnly.R
\name{get_BlueLake2018_FrontalCortexOnly}
\alias{get_BlueLake2018_FrontalCortexOnly}
\title{Get CellTypeDataset: ctd_BlueLake2018_FrontalCortexOnly}
\source{
\href{ https://pubmed.ncbi.nlm.nih.gov/29227469/}{Publication}

\href{https://www.nature.com/articles/s41467-019-11953-9}{Publication}

\code{
save_path <- file.path(
    "/Desktop/model_celltype_conservation/processed_data/EWCE",
    "standardized_CTD/BlueLake2018_FrontalCortexOnly.rds")
piggyback::pb_upload(file = save_path,
                     repo = "neurogenomics/phenomix",
                     overwrite = TRUE)
}
}
\usage{
get_BlueLake2018_FrontalCortexOnly()
}
\value{
\pkg{Seurat} object
}
\description{
Import the ctd_BlueLake2018_FrontalCortexOnly CellTypeDataset (CTD)
generated by \href{https://github.com/NathanSkene/EWCE}{EWCE}.
}
\details{
Contains mean expression (\emph{mean_exp}) matrices from sn/scRNA-seq
data collected across many tissues the human embryo.
Also contains precomputed \emph{specificity} and
\emph{specificity_quantiles} matrices.
}
