% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor.R
\name{get_cor}
\alias{get_cor}
\title{Extract correlation matrix}
\usage{
get_cor(
  obj,
  keys = NULL,
  assay = NULL,
  slot = NULL,
  graph_name = NULL,
  method = "pearson",
  return_obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Either a \pkg{Seurat} object or a feature*trait matrix.}

\item{keys}{The keys of reductions to extract from.}

\item{assay}{Assay to used to compute correlation graph 
if one does not already exist.}

\item{slot}{If \code{obj} is a \pkg{Seurat} object, which slot to extract
for correlations.}

\item{graph_name}{Name of the graph to extract.}

\item{method}{Pairwise correlation method.}

\item{return_obj}{Whether to return the \pkg{Seurat} object with a new
\code{obj@graphs} slot, or to simply return the sparse correlation matrix.}

\item{verbose}{Print messages.}
}
\value{
Trait-trait correlation matrix.
}
\description{
Extract correlation matrix stored as a graph.
}
\examples{
obj <- get_HPO()[seq(100),]
obj2 <- get_cor(obj = obj,
                keys = "pca")
}
