% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs2g.R
\name{get_cs2g}
\alias{get_cs2g}
\title{Import precomputed cS2G results}
\usage{
get_cs2g(
  dataset = c("finemapping_cS2G_UKBB", "gwas_catalog_cS2G"),
  save_dir = tools::R_user_dir(package = "phenomix", which = "cache"),
  as_matrix = FALSE,
  as_granges = FALSE,
  value_var = c("cS2G", "PIP"),
  min_value = 0,
  formula = "gene ~ ID",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{Dataset from the association 
\href{https://zenodo.org/record/7754032}{Zenodo} repository to use.}

\item{save_dir}{Directory to cache dataset in.}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{as_granges}{Convert the data object t0 
\link[GenomicRanges]{GRanges} format.}

\item{value_var}{Which column you want to fill the matrix with. 
Only used when \code{as_matrix==TRUE}.}

\item{min_value}{Minimum value of \code{value_var} to include in data.
Helpful when you're using a machine that can't handle large data.}

\item{formula}{A formula of the form LHS ~ RHS to cast, see Details.}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}
}
\value{
Named with two items: "data" and "obs" (observation metadata)
}
\description{
Precomputed results from the combined SNP-to-gene mapping model 
(\href{https://doi.org/10.1038/s41588-022-01087-y}{
Gazal et al., Nature Genetics, 2022}).\cr
}
\details{
\strong{gwas_catalog_cS2G}:\cr\cr
This compressed file contains cS2G score for 78,499 potentially causal
 SNP-gene-disease triplets from the NHGRI-EBI GWAS catalog
  (accessed file: gwas_catalog_v1.0-associations_e100_r2021-02-25.txt)\cr
PUBMEDID: PUBMED ID\cr
DISEASE.TRAIT: PHENOTYPE\cr
CHR: Chromosome\cr
POS: Position (hg19)\cr
SNP: SNP ID\cr
GENE: Gene ID\cr
cS2G: cS2G score\cr
INFO: S2G scores for the 10 main functional S2G
 (including the 7 cS2G constituent strategies)\cr

\strong{finemapping_cS2G_UKBB}:\cr\cr
Import 47 GWAS (from UK Biobank) mapped to genes using the
cS2G strategy.\cr\cr
This compressed file contains cS2G score for 138,716 potentially causal 
SNP-gene-disease triplets inferred from causal SNP-disease pairs with 
PIP>0.05 from functionally informed fine-mapping of 49
 UK Biobank diseases/traits.\cr
CHR: Chromosome\cr
BP: Position (hg19\cr
SNP: SNP ID\cr
DISEASE.TRAIT: PHENOTYPE\cr
PIP: Posterior inclusion probability\cr
GENE: Gene ID\cr
cS2G: cS2G score\cr
INFO: S2G scores for the 10 main functional S2G (including the 7 cS2G 
constituent strategies)\cr
This compressed file also contains a beta2 directory with PolyFun outputs
 for each traits (used in the omnigenicity analyses)\cr
}
\section{References}{

\href{https://doi.org/10.1101/2021.08.02.21261488}{
Gazal et al. 2021 preprint}
\href{https://zenodo.org/record/7754032}{Zenodo}
}

\examples{
meta <- get_cs2g_ukb()
}
