% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs2g_gwascatalog.R
\name{get_cs2g_gwascatalog}
\alias{get_cs2g_gwascatalog}
\title{Import precomputed cS2G results}
\usage{
get_cs2g_gwascatalog(
  URL = file.path("https://storage.googleapis.com", "broad-alkesgroup-public/cS2G",
    "gwas_catalog_cS2G", "gwas_catalog_v1.0-associations_e100_r2021-02-25.annot"),
  as_matrix = TRUE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{URL}{URL to cS2G dataset.}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{nThread}{Number of threads to use when reading in data.}

\item{verbose}{Print messages.}
}
\value{
Named with two items: "data" and "metadata"
}
\description{
Import 5,981 GWAS (from GWAS Catalog) mapped to genes using the
cS2G strategy.
}
\section{References}{


\href{https://doi.org/10.1101/2021.08.02.21261488}{
Gazal et al. 2021 preprint}

\href{https://alkesgroup.broadinstitute.org/cS2G}{
Broad Institute server}
}

