% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs2g_ukb.R
\name{get_cs2g_ukb}
\alias{get_cs2g_ukb}
\title{Import precomputed cS2G results}
\usage{
get_cs2g_ukb(
  URL = file.path("https://storage.googleapis.com", "broad-alkesgroup-public/cS2G",
    "finemapping_cS2G_UKBB", "finemapping_cS2G_UKBB_PIP05.annot"),
  as_matrix = TRUE,
  value_var = c("cS2G", "PIP"),
  min_value = 0,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{URL}{URL to cS2G dataset.}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{value_var}{Which column you want to fill the matrix with. 
Only used when \code{as_matrix==TRUE}.}

\item{min_value}{Minimum value of \code{value_var} to include in data.
Helpful when you're using a machine that can't handle large data.}

\item{nThread}{Number of threads to use when reading in data.}

\item{verbose}{Print messages.}
}
\value{
Named with two items: "data" and "metadata"
}
\description{
Import 47 GWAS (from UK Biobank) mapped to genes using the
cS2G strategy.
}
\section{References}{


\href{https://doi.org/10.1101/2021.08.02.21261488}{
Gazal et al. 2021 preprint}

\href{https://alkesgroup.broadinstitute.org/cS2G}{
Broad Institute server}
}

