% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd.R
\name{get_ctd}
\alias{get_ctd}
\title{Get a CellTypeDataset}
\usage{
get_ctd(fname = "BlueLake2018_FrontalCortexOnly.rds")
}
\value{
\pkg{CTD} object.
}
\description{
Get a CellTypeDataset object.
}
\examples{
ctd <- get_ctd()
}
