% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DEGAS.R
\name{get_DEGAS}
\alias{get_DEGAS}
\title{DeGAs seurat object}
\source{
\href{https://github.com/rivas-lab/public-resources/tree/master/uk_biobank/DeGAs}{DeGAs GitHub}

\href{https://www.nature.com/articles/s41467-019-11953-9}{Publication}

\code{
#### Embeddings/loadings
DEGAS_contributionGene <- readRDS(
    "/Desktop/phenome_decomposition/raw_data/DEGAS/contributionGene.rds")

#### Metadata
DEGAS_metadata <- read.csv(
    "/Desktop/phenome_decomposition/raw_data/DEGAS/metadata_processed.csv",
    check.names = FALSE)
colnames(DEGAS_metadata) <- gsub(" ","_",colnames(DEGAS_metadata))
colnames(DEGAS_metadata) <- gsub("[(]|[)]","",colnames(DEGAS_metadata))
rownames(DEGAS_metadata) <- DEGAS_metadata$label_phe_code

empty_mat <- Matrix::sparseMatrix(
    i = nrow(DEGAS_contributionGene@feature.loadings),
    j = nrow(DEGAS_contributionGene@cell.embeddings),
    dimnames = list(rownames(DEGAS_contributionGene@feature.loadings),
                  rownames(DEGAS_contributionGene@cell.embeddings)
                                  )
)
DEGAS_seurat <- Seurat::CreateSeuratObject(counts = empty_mat,
                                           assay = "genes", 
                                           meta.data = DEGAS_metadata)
DEGAS_seurat[["contributionGene"]]  <- DEGAS_contributionGene

save_path <- file.path(tempdir(),"DEGAS.rds")
saveRDS(DEGAS_seurat, save_path)
piggyback::pb_upload(file = save_path,
                     repo = "neurogenomics/phenomix",
                     overwrite = TRUE)
}
}
\usage{
get_DEGAS()
}
\value{
\pkg{Seurat} object
}
\description{
Contains embeddings/loadings for contributionGene,
as well as phenotype metadata.
}
\details{
\bold{NOTE:} The assay data is intentionally filled with
an empty sparse matrix due to size limits.
}
