% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DescartesHuman.R
\name{get_DescartesHuman}
\alias{get_DescartesHuman}
\title{Get CellTypeDataset: DescartesHuman}
\source{
\code{
save_path <- file.path("/Desktop/model_celltype_conservation/processed_data",
                       "EWCE/standardized_CTD/DescartesHuman.rds")
piggyback::pb_upload(file = save_path,
                     repo = "neurogenomics/phenomix",
                     overwrite = TRUE)
}
}
\usage{
get_DescartesHuman()
}
\value{
\pkg{Seurat} object
}
\description{
Import the DescartesHuman CellTypeDataset (CTD) generated by
\href{https://github.com/NathanSkene/EWCE}{EWCE}.
}
\details{
Contains mean expression (\emph{mean_exp}) matrices from sn/scRNA-seq
data collected across many tissues the human embryo.
Also contains precomputed \emph{specificity} and
\emph{specificity_quantiles} matrices.
}
