% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dPRS.R
\name{get_dPRS}
\alias{get_dPRS}
\title{dPRS seurat object}
\source{
\href{https://biobankengine.stanford.edu/downloads}{Global Biobank Engine}

\href{https://github.com/rivas-lab/public-resources/tree/master/uk_biobank/DeGAs}{DeGAs GitHub}

\href{https://www.nature.com/articles/s41431-021-00813-0?proof=t}{Original publication}

\code{
#### Download model weights (~0.7GB) ####
storage_dir <-
    file.path("/home/rstudio/raid1/bschilder/projects",
              "phenome_decomposition/raw_data/dPRS")
## Download model weights 
npz_file <- file.path(storage_dir,
                      "all_beta_center_p1e-06_20200506_500PCs.npz")
if(!file.exists(npz_file)){
    download.file(
        file.path("https://biobankengine.stanford.edu/static",
                  "degas-risk/all_beta_center_p1e-06_20200506_500PCs.npz"))
}
np <- reticulate::import("numpy")
dprs <- np$load(file = npz_file, allow_pickle = TRUE)

## Extract metadata 
meta_file <- file.path(storage_dir,"degas_n_977_traits.tsv")
if(!file.exists(meta_file)){
    download.file(
        file.path("https://biobankengine.stanford.edu/static",
                  "degas-risk/degas_n_977_traits.tsv"),
        meta_file
    )
}

meta <- data.table::fread(meta_file, data.table = FALSE)
colnames(meta) <- gsub("#","",colnames(meta))
meta <- meta |> dplyr::mutate(GBE_ID=gsub("_","-",GBE_ID))  
rownames(meta) <- meta$GBE_ID

## Reconstruct matrix
M_list <- phenomix::reconstruct_matrix(u_name = "V",
                                       v_name = "U",
                                       d_name = "D", 
                                       npz_obj = dprs,
                                       gene_level = FALSE,  
                                       translate_snp_positions = FALSE)
DR.contributionVar <- phenomix::: create_DimReducObject(
    reconstruction = M_list,
    assay = "variants",
    key = "contributionVar")
#### Variants assay ####
seurat <- Seurat::CreateSeuratObject(
    counts = Matrix::t(M_list$M2),
    meta.data = meta[rownames(M_list$M2),],
    assay = "variants")
#### DimReductions ####
seurat[["contributionVar"]] <- DR.contributionVar 
### Save local
save_path <- file.path(storage_dir,"dPRS.rds")
saveRDS(seurat,save_path)
### Upload remote
piggyback::pb_upload(file = save_path,
                     repo = "neurogenomics/phenomix",
                     overwrite = TRUE)
}
}
\usage{
get_dPRS()
}
\value{
\pkg{Seurat} object
}
\description{
Polygenic Risk Scores (PRS) applied to 
Decomposition of Genetic Associations (DeGAs) components. 
The dPRS methodology was first described in 
\href{https://www.nature.com/articles/s41431-021-00813-0?proof=t}{
Aguirre et al. (2021, European Journal of Human Genetics)}, where they used 
DeGAs to reduce ### GWAS traits to 500 components.
}
\details{
Specifically, we used the
 \emph{DeGAs model weights: all_beta_center_p1e-06_20200506_500PCs.npz} file 
 found on the 
 \href{https://biobankengine.stanford.edu/downloads}{
 Global Biobank Engine}.
 Please see \link[phenomix]{get_DEGAS} for more details on 
the DeGAs methodology.
}
\examples{
dprs <- phenomix::get_dPRS()
}
