% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HPO.R
\name{get_HPO}
\alias{get_HPO}
\title{Get the Human Phenotype Ontology}
\source{
\code{
save_path <-  file.path("/Desktop/phenome_decomposition",
                        "raw_data/HPO/HPO_seurat.rds")
obj <- prepare_HPO(save_path = save_path)
piggyback::pb_upload(file = save_path, 
                     repo = "neurogenomics/phenomix",
                     overwrite = TRUE)
}
}
\usage{
get_HPO()
}
\value{
\pkg{Seurat} object
}
\description{
Import the Human Phenotype Ontology (HPO) as a Seurat object
with precalculated variable genes, specificity stored as an assay, 
and PCA/UMAP reductions.
}
