% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_genes.R
\name{get_mhc_genes}
\alias{get_mhc_genes}
\title{Get MHC genes}
\source{
\href{https://www.nature.com/articles/s41467-019-11953-9}{DeGAs}
}
\usage{
get_mhc_genes(
  seqnames = "chr6:25000000-34000000",
  ref_genome = "GRCh37",
  force_new = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seqnames}{Genomic ranges to search for genes.}

\item{ref_genome}{name of the reference genome used for the GWAS ("GRCh37" or
"GRCh38"). Argument is case-insensitive. Default is NULL which infers the
reference genome from the data.}

\item{force_new}{Query for all genes within \code{seqnames} using
\link[TxDb.Hsapiens.UCSC.hg19.knownGene]{TxDb.Hsapiens.UCSC.hg19.knownGene}
or
\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}
as a reference.
If \code{FALSE} \code{(DEFAULT)}, will instead
return a stored \link[data.table]{data.table}
aligned to the "GRCh37" reference genome.}

\item{verbose}{Print messages.}
}
\description{
Retrieve genes from the MHC
(Major Histocompatibility Complex) region.
}
\details{
The MHC region tends to be ubiquitously associated with all GWAS,
so removing it allows us to focus on non-MHC signals of interest.
For further information related to the MHC region, see the sources below.
}
\examples{
MHC_genes <- get_mhc_genes()
}
