% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology_groups.R
\name{get_ontology_groups}
\alias{get_ontology_groups}
\title{Assign ontology groups}
\usage{
get_ontology_groups(
  ontology,
  term_ids = NULL,
  depth_levels = seq(2, 4),
  verbose = TRUE
)
}
\arguments{
\item{ontology}{Ontology to search.}

\item{depth_levels}{Which depth level to assign categories with.}

\item{verbose}{Print messages.}
}
\value{
\code{data.table} with group assignments at each depth.
}
\description{
Searches for higher-level groups within the ontology
 (at depths 2-4 by default), and then assigns each term
  in the ontology to one of those groups.
 
If a given term matches with >1 group at a given depth, 
the term will be assigned to the first group arbitrarily.
}
