% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_sdreport}
\alias{get_sdreport}
\title{Get sdreport for predictions / coefficients, copying glmmTMB 'fast' implementation}
\usage{
get_sdreport(object)
}
\arguments{
\item{object}{The fitted phenomix model}
}
\description{
Get sdreport for predictions / coefficients, copying glmmTMB 'fast' implementation
}
\keyword{internal}
