% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_features.R
\name{get_top_features}
\alias{get_top_features}
\title{Plot the top feature loadings}
\usage{
get_top_features(
  obj,
  reduction = NULL,
  n_features = 3,
  n_features_plot = 20,
  factors = seq(1, 10),
  invert_vars = FALSE,
  show_plot = TRUE,
  title = NULL,
  return_melted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{\pkg{Seurat} object or dimensionality reduction object.}

\item{reduction}{Which reduction to extract from. 
Defaults to first available reduction. 
Only used if \code{obj} is of class \code{Seurat}.}

\item{n_features}{Number of top features per factor to select.}

\item{n_features_plot}{Number of top features to plot.}

\item{factors}{Which factors to plot.}

\item{invert_vars}{Switch the axes of factors and loadings in plot.}

\item{show_plot}{Whether to print the plot or simply return it.}

\item{title}{Plot title.}

\item{verbose}{Print messages.}
}
\value{
top_features \code{data.table}
}
\description{
Plot the top feature loadings for from a given reduction.
}
\examples{
data("DEGAS_seurat")
top_features <- get_top_features(obj=DEGAS_seurat)
}
