% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_phenos.R, R/get_top_phenotypes.R
\name{get_top_phenotypes}
\alias{get_top_phenotypes}
\title{Plot the top phenotypes}
\usage{
get_top_phenotypes(
  obj,
  metadata = NULL,
  reduction = NULL,
  n_phenotypes = 3,
  factors_plot = 1:10,
  invert_vars = FALSE,
  show_plot = TRUE,
  title = NULL,
  x = "phenotype",
  y = "loading",
  verbose = TRUE
)

get_top_phenotypes(
  obj,
  metadata = NULL,
  reduction = NULL,
  n_phenotypes = 3,
  factors_plot = 1:10,
  invert_vars = FALSE,
  show_plot = TRUE,
  title = NULL,
  x = "phenotype",
  y = "loading",
  verbose = TRUE
)
}
\arguments{
\item{obj}{\pkg{Seurat} object or dimensionality reduction object.}

\item{metadata}{Phenotype metadata.
Not needed if \code{obj} is a \pkg{Seurat} object.}

\item{reduction}{Reduction to use. If \code{NULL}, defaults to first available reduction.}

\item{n_phenotypes}{Number of top phenotypes per reduction factor to select.}

\item{invert_vars}{Switch the axes of factors and loadings in plot.}

\item{show_plot}{Whether to print the plot or simply return it.}

\item{title}{Plot title.}

\item{x}{x-axis variable.}

\item{y}{y-axis variable.}

\item{verbose}{Print messages.}
}
\value{
\code{data.table} of top phenotypes

\code{data.table} of top phenotypes
}
\description{
Create a radar chart of the phenotypes with the
highest loadings for each reduction factor.

Create a radar chart of the phenotypes with the
highest loadings for each reduction factor.
}
\examples{
degas <- get_DEGAS()
top_phenos <- get_top_phenotypes(obj = degas)
degas <- get_DEGAS()
top_phenos <- get_top_phenotypes(obj = degas)
}
