% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_matrix_magma.R
\name{gwas_matrix_magma}
\alias{gwas_matrix_magma}
\title{Construct gene x trait matrix from MAGMA results}
\usage{
gwas_matrix_magma(
  magma_files,
  metric = c("ADJ_ZSTAT", "ZSTAT", "P"),
  drop_MHC = TRUE,
  save_path = NULL,
  nCores = 1,
  fillna = TRUE,
  agg_FUN = "mean",
  verbose = TRUE
)
}
\arguments{
\item{magma_files}{Paths to \emph{genes.out} files
produced by \link[MAGMA.Celltyping]{map.snps.to.genes}.}

\item{metric}{Which metric to fill the matrix with
\itemize{
 \item{\code{"ADJ_ZSTAT"} : }{MAGMA Z-statistic after adjusting for
 gene length and other confounds (\emph{DEFAULT}).}
 \item{\code{"ZSTAT"} : }{Unadjusted MAGMA Z-statistic.}
 \item{\code{"P"} : }{MAGMA P-value.}
}}

\item{drop_MHC}{Drop genes from the MHC
(Major Histocompatibility Complex) region.}

\item{save_path}{Path to save results as RDS object.}

\item{nCores}{Number of cores to parallelise across.}

\item{fillna}{Value to fill \code{NA}s with.}

\item{agg_FUN}{Aggregation function.}

\item{verbose}{Print messages.}
}
\value{
Sparse matrix
}
\description{
Once you've run \link[MAGMA.Celltyping]{map.snps.to.genes} on multiple GWAS,
you can aggregate the gene-wise results into a single gene x trait matrix.
}
