% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_abc.R
\name{import_abc}
\alias{import_abc}
\title{Import ABC model predictions}
\usage{
import_abc(dataset = "Nasser2020", abc = NULL, nCores = 1, verbose = TRUE)
}
\arguments{
\item{dataset}{Which ABC model to import.}

\item{abc}{Use a previously downloaded ABC \link[data.table]{data.table}.}

\item{nCores}{Number of cores to parallelise across.}

\item{verbose}{Print messages.}

\item{method}{Method to use for mapping SNPs to genes.}
}
\value{
\code{gene_hits} \link[data.table]{data.table}
}
\description{
Import Activity-By-Contact (ABC) model predictions 
from multiple publications.
}
\section{References}{


\bold{Nasser2020}\cr
\href{https://www.nature.com/articles/s41586-021-03446-x}{Nasser et al. 2020, Nature}

\bold{Fulco2019}\cr
\href{https://www.nature.com/articles/s41588-019-0538-0?proof=t}{Fulco et al. 2019, Nature}
\href{https://github.com/broadinstitute/ABC-Enhancer-Gene-Prediction}{Fulco et al. 2019, GitHub}

\href{https://www.engreitzlab.org/resources/}{ABC models from multiple Engreitz Lab publications}
}

