% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_genome_builds.R
\name{infer_genome_builds}
\alias{infer_genome_builds}
\title{Infer genome builds}
\usage{
infer_genome_builds(
  sumstats,
  header_only = TRUE,
  sampled_snps = 10000,
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{sumstats}{data table/data frame obj of the summary statistics file for
the GWAS ,or file path to summary statistics file.}

\item{sampled_snps}{Downsample the number of SNPs used when inferring genome
build to save time.}

\item{nCores}{Number of threads to use for parallel processes.}

\item{verbose}{Print messages.}
}
\value{
ref_genome the genome build of the data
}
\description{
Infers the genome build of  summary statistics files (GRCh37 or GRCh38) 
from the data. Uses SNP (RSID) & CHR & BP to get genome build.
}
\examples{
#Pass path to Educational Attainment Okbay sumstat file to a temp directory

eduAttainOkbayPth <- system.file("extdata","eduAttainOkbay.txt", 
                                 package="MungeSumstats")

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks
is_32bit_windows <- 
.Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
ref_genome <- infer_genome_builds(sumstats=eduAttainOkbayPth)                                  
}
}
