% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_gsea.R
\name{iterate_gsea}
\alias{iterate_gsea}
\title{Iterate GSEA}
\usage{
iterate_gsea(
  xmat,
  ymat,
  correction_method = "BH",
  qvalue_thresh = 0.05,
  x_quantiles = 10,
  y_quantiles = 10,
  use_quantiles = 10,
  nCores = 1
)
}
\arguments{
\item{xmat}{gene x trait matrix.}

\item{ymat}{gene x celltype matrix.}

\item{correction_method}{Multiple-testing correction method 
to be passed to \code{stats::p.adjust}.}

\item{qvalue_thresh}{q.value threshold to use when report 
significant results summary.}

\item{x_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{y_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{use_quantiles}{Which quantiles in to use in 
\code{xmat} and \code{ymat}.}

\item{nCores}{Number of cores to use in parallel. 
Will optimize if \code{NULL}.}
}
\value{
\code{data.table} of enrichment results.
}
\description{
Iterate gene set enrichment analysis across all traits and celltypes.
}
\examples{
### DeGAs loadings
obj <- get_DEGAS()
xmat <- get_varm(obj = obj)
xmat <- xmat[, 1:10] # Let's use just 10 components as an example

### Celltype Dataset
ctd <- get_BlueLake2018_FrontalCortexOnly()
ymat <- ctd[[1]]$specificity
res_gsea <- iterate_gsea(xmat = xmat,
                         ymat = ymat,
                         nCores = 1)
}
