% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_char_fun.R
\name{map_id_sep}
\alias{map_id_sep}
\title{Map ID separators}
\usage{
map_id_sep(nms, replace_char = list(. = ":", `_` = ":"))
}
\description{
Replace the separator in an ID in order to standardise it.
}
\examples{
map_id_sep(c("HP:0001","HP.0001","HP_0001"))
}
