% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_phenotypes.R
\name{map_phenotypes}
\alias{map_phenotypes}
\title{Map phenotypes}
\usage{
map_phenotypes(phenotypes, ont = HPOExplorer::get_hpo())
}
\arguments{
\item{phenotypes}{A character vector of phenotype terms.}

\item{ont}{An ontology of class \code{ontology_index}.}
}
\value{
Named vector mapping the phenotypes onto an ontology.
}
\description{
Map phenotypes onto a standardised ontology.
}
\examples{
meta <- MungeSumstats::find_sumstats()
phenotypes <- unique(meta$trait)
pmap <- map_phenotypes(phenotypes=phenotypes) 
}
