% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snps2genes.R
\name{map_snps2genes}
\alias{map_snps2genes}
\title{Map SNPs to genes}
\usage{
map_snps2genes(
  dat,
  method = c("txdb", "abc"),
  dataset = "Nasser2020",
  abc = NULL,
  agg_var = "SYMBOL",
  drop_na = TRUE,
  adjust_z = TRUE,
  drop_mhc = TRUE,
  model = NULL,
  log_vars = c("NSNPS", "NPARAM", "GENELEN"),
  formula = ZSTAT ~ NSNPS + logNSNPS + NPARAM + logNPARAM + GENELEN + logGENELEN,
  nCores = 1,
  save_path = tempfile(),
  return_path = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics munged by
\link[MungeSumstats]{format_sumstats}.
Can be a path to the saved file or \link[data.table]{data.table}.}

\item{method}{Method to use for mapping SNPs to genes.}

\item{dataset}{Which ABC model to import.}

\item{abc}{Use a previously downloaded ABC \link[data.table]{data.table}.}

\item{agg_var}{Variable to aggregate \code{merged_hits} by.}

\item{drop_na}{Drop rows with \code{NA} in any column.}

\item{adjust_z}{Whether to adjust Z-statistic using
the \code{model} and \code{formula}.}

\item{drop_mhc}{Drop genes from the MHC
(Major Histocompatibility Complex) region.}

\item{model}{Statistical model to use.
Defaults to \link[stats]{lm}.}

\item{log_vars}{Variables to perform natural log transformation on first.
Only run on variables available in \code{dat}.}

\item{formula}{Formula to use in \code{model}.}

\item{nCores}{Number of cores to parallelise across.}

\item{save_path}{Path to save results to.}

\item{return_path}{Whether to return the path to the saved results.}

\item{verbose}{Print messages.}

\item{agg_by_genes}{Aggregate results by genes (\code{TRUE}) or SNPs
(\code{FALSE}).}
}
\value{
gene_hits \link[data.table]{data.table}.
}
\description{
Map SNPs (both coding and non-coding) onto genes
using different methods.
}
\section{Other packages worth exploring}{

\href{https://github.com/thewonlab/H-MAGMA}{H-MAGMA}
Really hard to use with no docs on how to create new annot files.
It's literally just MAGMA (when you provide some extra args).

\href{https://www.rdocumentation.org/packages/PAGWAS/versions/2.0/topics/snps.to.genes}{PAGWAS}

\href{https://cran.r-project.org/web/packages/rsnps/rsnps.pdf}{rsnp}

\href{orthogene}{https://github.com/neurogenomics/orthogene}
Protein-coding variants only?

\href{OpenTargets}{https://genetics.opentargets.org/}
Could use colocalization results, but these are both GWAS- and QTL-specific.
Could compute mean coloc score per gene for each SNP and use that.
Other position-based functional data might still be useful.
}

\examples{
dat <- MungeSumstats::formatted_example()
dat2 <- map_snps2genes(dat)
}
