% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snps2genes_abc.R
\name{map_snps2genes_abc}
\alias{map_snps2genes_abc}
\title{Map SNPs to genes: ABC}
\usage{
map_snps2genes_abc(
  dat,
  ref_genome = "GRCh37",
  dataset = "Nasser2020",
  abc = NULL,
  agg_var = "SYMBOL",
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics munged by
\link[MungeSumstats]{format_sumstats}.
Can be a path to the saved file or \link[data.table]{data.table}.}

\item{ref_genome}{name of the reference genome used for the GWAS ("GRCh37" or
"GRCh38"). Argument is case-insensitive. Default is NULL which infers the
reference genome from the data.}

\item{dataset}{Which ABC model to import.}

\item{abc}{Use a previously downloaded ABC \link[data.table]{data.table}.}

\item{agg_var}{Variable to aggregate \code{merged_hits} by.}

\item{nCores}{Number of cores to parallelise across.}

\item{verbose}{Print messages.}

\item{method}{Method to use for mapping SNPs to genes.}
}
\value{
\code{gene_hits} \link[data.table]{data.table}
}
\description{
Map SNPs (both coding and non-coding) onto genes
using an Activity-by-Contact (ABC) model.
}
\section{References}{

\href{https://www.nature.com/articles/s41586-021-03446-x}{Nasser et al. 2020, Nature}

\href{https://www.nature.com/articles/s41588-019-0538-0?proof=t}{Fulco et al. 2019, Nature}
\href{https://github.com/broadinstitute/ABC-Enhancer-Gene-Prediction}{Fulco et al. 2019, GitHub}

\href{https://www.engreitzlab.org/resources/}{ABC models from multiple Engreitz Lab publications}
}

\keyword{internal}
