% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snps2genes_cs2g.R
\name{map_snps2genes_cs2g}
\alias{map_snps2genes_cs2g}
\title{Map SNPs to genes: cS2G}
\usage{
map_snps2genes_cs2g(
  sumstats,
  ref_genome = "GRCh37",
  dataset = "Nasser2020",
  abc = NULL,
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{nCores}{Number of cores to parallelise across.}

\item{verbose}{Print messages.}

\item{sumstats_file}{GWAS summary statistics munged by
\link[MungeSumstats]{format_sumstats}.
Can be a path to the saved file or \link[data.table]{data.table}.}

\item{method}{Method to use for mapping SNPs to genes.}
}
\value{
\code{gene_hits} \link[data.table]{data.table}
}
\description{
Map SNPs (both coding and non-coding) onto genes
using the combined SNP-to-gene (cS2G) strategy..
}
\section{References}{


\href{https://doi.org/10.1101/2021.08.02.21261488}{
Gazal et al. 2021 preprint}

\href{https://alkesgroup.broadinstitute.org/cS2G}{
Broad Institute server}
}

