% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snps2genes_txdb.R
\name{map_snps2genes_txdb}
\alias{map_snps2genes_txdb}
\title{Map SNPs to genes: txdb}
\usage{
map_snps2genes_txdb(
  dat,
  promoter_upstream = 35000,
  promoter_downstream = 10000,
  queries = c("promoter", "coding", "intron", "threeUTR", "splicesite"),
  drop_na = TRUE,
  agg_var = "SYMBOL",
  zscore_col = "BETA_mean",
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics munged by
\link[MungeSumstats]{format_sumstats}.
Can be a path to the saved file or \link[data.table]{data.table}.}

\item{promoter_upstream}{How many basepairs upstream of known promoters
to search for genes.}

\item{promoter_downstream}{How many basepairs dowmstream of known promoters
to search for genes.}

\item{queries}{Genomic region types to search for overlap with \code{gr}.
See \link[VariantAnnotation]{AllVariants} for details.}

\item{drop_na}{Drop rows with \code{NA} in any column.}

\item{agg_var}{Variable to aggregate \code{merged_hits} by.}

\item{nCores}{Number of cores to parellelise across.}

\item{verbose}{Print messages.}

\item{method}{Method to use for mapping SNPs to genes.}
}
\value{
\code{gene_hits} \link[data.table]{data.table}
}
\description{
Map SNPs (both coding and non-coding) onto genes
using
\link[TxDb.Hsapiens.UCSC.hg19.knownGene]{TxDb.Hsapiens.UCSC.hg19.knownGene}
or
\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}
as a reference.
}
\keyword{internal}
