% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snps_txdb.R
\name{map_snps_txdb}
\alias{map_snps_txdb}
\title{Map SNPs to genes: txdb}
\usage{
map_snps_txdb(
  dat,
  promoter_upstream = 35000,
  promoter_downstream = 10000,
  queries = c("promoter", "coding", "intron", "threeUTR", "splicesite"),
  all_variants = FALSE,
  ref_genome = "GRCh37",
  return_merged = TRUE,
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics munged by
\link[MungeSumstats]{format_sumstats}.}

\item{promoter_upstream}{How many basepairs upstream of known promoters
to search for genes.}

\item{promoter_downstream}{How many basepairs dowmstream of known promoters
to search for genes.}

\item{queries}{Genomic region types to search for overlap with \code{gr}.
See \link[VariantAnnotation]{AllVariants} for details.}

\item{all_variants}{Search all genomic regions regardless of type.
This can take longer and returns less targeted results.}

\item{ref_genome}{name of the reference genome used for the GWAS ("GRCh37" or
"GRCh38"). Argument is case-insensitive. Default is NULL which infers the
reference genome from the data.}

\item{verbose}{Print messages.}
}
\value{
\code{hits} \link[data.table]{data.table}
}
\description{
Map both coding and non-coding SNPs to genes using
#' \link[TxDb.Hsapiens.UCSC.hg19.knownGene]{TxDb.Hsapiens.UCSC.hg19.knownGene}
or
\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}
as a reference.
}
\details{
Permits selection of different region types with
\code{queries} argument.
}
\keyword{internal}
