% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MHC_genes}
\alias{MHC_genes}
\title{MHC genes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 805 rows and 16 columns.
}
\source{
\href{https://www.nature.com/articles/s41467-019-11953-9}{DeGAs}
 
\code{
MHC_genes <- get_mhc_genes()
usethis::use_data(MHC_genes, overwrite = TRUE)
}
}
\usage{
MHC_genes
}
\description{
Retrieve genes from the MHC 
(Major Histocompatibility Complex) region.
}
\details{
The MHC region tends to be ubiquitously associated with all GWAS,
so removing it allows us to focus on non-MHC signals of interest.
For further information related to the MHC region, see the sources below.
}
\keyword{datasets}
