% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengwas_meta.R
\name{opengwas_meta}
\alias{opengwas_meta}
\title{OpenGWAS metadata}
\usage{
opengwas_meta(
  meta = phenomix::OpenGWAS,
  ids = NULL,
  host = paste0("https://phenomix.dsi.ic.ac.uk/", "MAGMA_Files_Public/data/GWAS_munged/"),
  verbose = TRUE
)
}
\arguments{
\item{meta}{Metadata for OpenGWAS datasets.}

\item{ids}{A list of OpenGWAS dataset IDs.}

\item{host}{phenomix database host URL.}

\item{verbose}{Print messages.}
}
\value{
data.table
}
\description{
Prepare metadata for GWAS/QTL datasets from the OpenGWAS database.
}
\examples{
meta <- opengwas_meta() 
}
