% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pars}
\alias{pars}
\title{Get parameters from model object, copying glmmTMB 'fast' implementation}
\usage{
pars(object)
}
\arguments{
\item{object}{The fitted phenomix model}
}
\description{
Get parameters from model object, copying glmmTMB 'fast' implementation
}
\examples{
data(fishdist)

# example of fitting fixed effects, no trends, no random effects
set.seed(1)
datalist <- create_data(fishdist[which(fishdist$year > 1970), ],
  asymmetric_model = FALSE,
  est_mu_re = FALSE, est_sigma_re = FALSE
)
fit <- fit(datalist)
p <- pars(fit)
names(p)
}
