% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomix-package.R
\docType{package}
\name{phenomix-package}
\alias{phenomix}
\alias{phenomix-package}
\title{The 'phenomix' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\seealso{
Useful links:
\itemize{
  \item \url{https://noaa-nwfsc.github.io/phenomix}
  \item \url{https://github.com/noaa-nwfsc/phenomix}
}

}
\author{
\strong{Maintainer}: Eric J. Ward \email{eric.ward@noaa.gov}

Other contributors:
\itemize{
  \item Samantha M. Wilson [contributor]
  \item Joseph H. Anderson [contributor]
}

}
\keyword{internal}
