% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomix_merge.R
\name{phenomix_merge}
\alias{phenomix_merge}
\title{Merge phenomix queries}
\usage{
phenomix_merge(
  res,
  value_var = "BETA",
  key_var = c("SNP", "A1", "A2", "CHR", "BP"),
  save_path = NULL,
  as_matrix = TRUE,
  all = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{res}{Results from \link[phenomix]{query_phenomix}. 
Can be a list of \link[data.table]{data.table}s or sparse matrices.}

\item{value_var}{Column to fill the matrix with.}

\item{key_var}{A column (or set of columns) to be used when
merging datasets together.}

\item{save_path}{Path to save matrix to in RDS format
(set to \code{NULL} to skip this step).}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{all}{logical; \code{all = TRUE} is shorthand to save setting both
\code{all.x = TRUE} and \code{all.y = TRUE}.}

\item{verbose}{Print messages.}
}
\value{
Sparse matrix
}
\description{
Merge results from \link[phenomix]{query_phenomix} 
into a single sparse matrix or \link[data.table]{data.table} filled with summary statistics.
}
\examples{
query_granges <- GenomicRanges::GRanges(c("2:15000-17000","3:60000-70000"))
ids <- tail(phenomix::OpenGWAS[!grepl("^finn",id),]$id,3)
res <- phenomix_query(ids=ids, query_granges=query_granges)
X <- phenomix_merge(res) 
}
