% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomix_query.R
\name{phenomix_query}
\alias{phenomix_query}
\title{Query \href{https://phenomix.dsi.ic.ac.uk/}{phenomix}}
\usage{
phenomix_query(
  ids = NULL,
  meta = opengwas_meta(ids = ids),
  save_dir = tempdir(),
  query_granges = NULL,
  data_type = c("sumstats", "magma"),
  as_matrix = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ids}{A list of OpenGWAS dataset IDs.}

\item{meta}{Metadata for OpenGWAS datasets.}

\item{save_dir}{Directory to save query results to.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{data_type}{Data type to query:
\itemize{
\item{"sumstats" : }{Genome-wide SNP-level summary statistics 
that have previously been harmonised using 
\link[MungeSumstats]{format_sumstats}.}
\item{"magma" : }{Genome-wide gene-level 
\href{https://ctg.cncr.nl/software/magm}{MAGMA} results.}
}}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[echotabix:query_table]{echotabix::query_table}}
  \describe{
    \item{\code{target_path}}{Path to tabix file.}
    \item{\code{target_index}}{Tabix index file for \code{target_path}.}
    \item{\code{query_method}}{Method used for querying. 
See \link[echotabix]{query} for available options.}
    \item{\code{local}}{Whether \code{target_path} is stored locally or 
on a remote server/website.
By default (\code{NULL}) will infer local status and 
use the appropriate \code{query_method}.}
    \item{\code{overlapping_only}}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}
    \item{\code{query_save}}{Whether to save the queried data subset.}
    \item{\code{save_path}}{File path to save query subset to (as table).}
    \item{\code{cleanup_tbi}}{Remove local copies of tabix index file (\emph{.tbi}) 
after completing queries.}
    \item{\code{conda_env}}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}
    \item{\code{nThread}}{Number of threads to use.}
  }}
}
\value{
data.table of genomic regions
}
\description{
Query the \href{https://phenomix.dsi.ic.ac.uk/}{phenomix} database
 for harmonised GWAS or QTL datasets.
}
\examples{
query_granges <- GenomicRanges::GRanges(c("2:15000-16000","3:60000-61000"))
res <- phenomix_query(ids="bbj-a-1", query_granges=query_granges)
}
