% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomix_query_batched.R
\name{phenomix_query_batched}
\alias{phenomix_query_batched}
\title{Query \href{https://phenomix.dsi.ic.ac.uk/}{phenomix}}
\usage{
phenomix_query_batched(
  ids = NULL,
  meta = opengwas_meta(ids = ids),
  save_dir = tempdir(),
  query_granges = NULL,
  batch_size = 100,
  as_matrix = TRUE,
  value_var = "BETA",
  key_var = c("SNP", "A1", "A2", "CHR", "BP"),
  all = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{ids}{A list of OpenGWAS dataset IDs.}

\item{meta}{Metadata for OpenGWAS datasets.}

\item{save_dir}{Directory to save query results to.}

\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{batch_size}{Number of datasets to process per batch.}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{value_var}{Column to fill the matrix with.}

\item{key_var}{A column (or set of columns) to be used when
merging datasets together.}

\item{all}{logical; \code{all = TRUE} is shorthand to save setting both
\code{all.x = TRUE} and \code{all.y = TRUE}.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[echotabix:query_table]{echotabix::query_table}}
  \describe{
    \item{\code{target_path}}{Path to tabix file.}
    \item{\code{target_index}}{Tabix index file for \code{target_path}.}
    \item{\code{query_method}}{Method used for querying. 
See \link[echotabix]{query} for available options.}
    \item{\code{local}}{Whether \code{target_path} is stored locally or 
on a remote server/website.
By default (\code{NULL}) will infer local status and 
use the appropriate \code{query_method}.}
    \item{\code{overlapping_only}}{Remove variants that do not overlap with the 
positions in \code{query_dat}.}
    \item{\code{query_save}}{Whether to save the queried data subset.}
    \item{\code{save_path}}{File path to save query subset to (as table).}
    \item{\code{cleanup_tbi}}{Remove local copies of tabix index file (\emph{.tbi}) 
after completing queries.}
    \item{\code{conda_env}}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}
    \item{\code{nThread}}{Number of threads to use.}
  }}
}
\value{
data.table of genomic regions
}
\description{
Query the \href{https://phenomix.dsi.ic.ac.uk/}{phenomix} database
 for harmonised GWAS or QTL datasets.
}
\examples{
query_granges <- GenomicRanges::GRanges(c("2:15000-17000","3:60000-70000"))
ids <- tail(phenomix::OpenGWAS[!grepl("^finn",id),]$id,6)
X <- phenomix_query_batched(ids=ids,
                            query_granges=query_granges,
                            batch_size=2)
}
