% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag_enrich.R
\name{plot_dag_enrich}
\alias{plot_dag_enrich}
\title{Plot enrichment of terms in a DAG ontology}
\usage{
plot_dag_enrich(
  obj,
  ont,
  min_hits = 3,
  min_offspring = 100,
  min_depth = NULL,
  exact_depth = NULL,
  exact_ont_ancestors = FALSE,
  sort_by = c(log2_fold_enrichment = -1),
  cluster_col = "seurat_clusters",
  id_col = NULL,
  label_col = "dag_enrich.name_wrap",
  cluster_id_alpha = 0.9,
  cluster_id_size = 3,
  color_col = cluster_col,
  replace_char = list(. = ":", `_` = ":"),
  p_threshold = 0.05,
  q_threshold = 0.05,
  preferred_palettes = "kovesi.cyclic_mygbm_30_95_c78",
  add_labels = TRUE,
  reduction = NULL,
  label_alpha = 0.5,
  point_alpha = 0.4,
  point_size = 0.7,
  label_size = 3,
  label_width = 20,
  repel = TRUE,
  box = TRUE,
  box.padding = 1,
  max.overlaps = 100,
  na.value = "grey50",
  force_new = FALSE,
  return_obj = FALSE,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{repel}{Use \code{geom_text_repel} to create nicely-repelled labels}

\item{box}{Use geom_label/geom_label_repel (includes a box around the text
labels)}

\item{box.padding}{Amount of padding around bounding box, as unit or number.
Defaults to 0.25. (Default unit is lines, but other units can be specified
by passing \code{unit(x, "units")}).}

\item{max.overlaps}{Exclude text labels when they overlap too many other
things. For each text label, we count how many other text labels or other
data points it overlaps, and exclude the text label if it has too many overlaps.
Defaults to 10.}

\item{...}{
  Arguments passed on to \code{\link[Seurat:LabelClusters]{Seurat::LabelClusters}}
  \describe{
    \item{\code{plot}}{A ggplot2-based scatter plot}
    \item{\code{id}}{Name of variable used for coloring scatter plot}
    \item{\code{clusters}}{Vector of cluster ids to label}
    \item{\code{labels}}{Custom labels for the clusters}
    \item{\code{split.by}}{Split labels by some grouping label, useful when using
\code{\link[ggplot2]{facet_wrap}} or \code{\link[ggplot2]{facet_grid}}}
    \item{\code{geom}}{Name of geom to get X/Y aesthetic names for}
    \item{\code{position}}{How to place the label if repel = FALSE. If "median", place
the label at the median position. If "nearest" place the label at the
position of the nearest data point to the median.}
  }}
}
\description{
Plot enrichment of terms in a DAG ontology.
}
\examples{
ont <- KGExplorer::get_ontology("hp")
obj <- get_HPO()
out <- plot_dag_enrich(obj, ont)
}
