% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot enrichment results}
\usage{
plot_enrichment(
  res,
  types = c("heat", "bar"),
  qvalue_thresh = 0.05,
  top_n = 50,
  show_plot = TRUE,
  save_path = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{res}{Enrichment results table.}

\item{types}{A list of plot types to create}

\item{qvalue_thresh}{Threshold to filter results.}

\item{top_n}{Number of top results to plot
(sorted by smallest p-values).}

\item{show_plot}{Print the plot.}

\item{save_path}{Path to save plot to.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[heatmaply]{heatmaply}.}
}
\value{
A list of ggplot and/or plotly objects.
}
\description{
Plot results from \link[phenomix]{iterate_lm} or
\link[phenomix]{iterate_gsea}.
}
