% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_heat.R
\name{plot_enrichment_heat}
\alias{plot_enrichment_heat}
\title{Plot enrichment results: heatmap}
\usage{
plot_enrichment_heat(
  res,
  qvalue_thresh = 0.05,
  show_plot = TRUE,
  save_path = NULL,
  verbose = TRUE,
  ...
)
}
\description{
Plot enrichment results: heatmap
}
\keyword{internal}
