% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors_sankey.R
\name{plot_factors_sankey}
\alias{plot_factors_sankey}
\title{Plot factors: Sankey}
\usage{
plot_factors_sankey(
  factor.traits,
  factor.celltypes,
  varm = NULL,
  max_factors = NULL,
  q_threshold = 0.05,
  logfc_threshold = NULL,
  value_var = "q",
  drop = TRUE,
  flow.alpha = 0.8,
  label_size = 3,
  label_color = "white",
  label_fill = ggplot2::alpha("black", 0.7),
  xtext_size = 12,
  show.legend = FALSE,
  ...
)
}
\arguments{
\item{show.legend}{show legend in plot}

\item{...}{
  Arguments passed on to \code{\link[ggsankey:geom_sankey]{ggsankey::geom_sankey}}
  \describe{
    \item{\code{mapping}}{provide you own mapping. both x and y need to be numeric.}
    \item{\code{data}}{provide you own data}
    \item{\code{position}}{change position}
    \item{\code{na.rm}}{remove missing values}
    \item{\code{space}}{space between nodes in the y-direction}
    \item{\code{type}}{either 'sankey' or 'alluvial'}
    \item{\code{width}}{width of nodes}
    \item{\code{smooth}}{how much smooth should the curve have? More means steeper curve.}
    \item{\code{inherit.aes}}{should the geom inherit aesthetics}
  }}
}
\description{
Create a Sankey plot from enrichment results performed on the 
feature weights and sample loadings from a factorization model.
}
