% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_density.R
\name{plot_feature_density}
\alias{plot_feature_density}
\title{Plot feature density}
\usage{
plot_feature_density(
  obj,
  features = sample(rownames(obj), 3),
  alpha = 0.5,
  method = c("wkde", "ks")[1],
  joint = TRUE,
  joint_only = FALSE,
  joint_title = if (length(unique(features)) > 10) {
     paste0("Density of ",
    length(unique(features)), " features")
 },
  axis_titles = "collect",
  axes = "collect",
  guides = NULL,
  ...
)
}
\arguments{
\item{features}{Features (e.g. genes) to visualize}

\item{method}{Kernel density estimation method:
\itemize{
\item \code{ks}: Computes density using the \code{kde} function from the
 \code{ks} package.
\item \code{wkde}: Computes density using a modified version of the
\code{kde2d} function from the \code{MASS}
package to allow weights. Bandwidth selection from the \code{ks} package
 is used instead.
}}

\item{joint}{Return joint density plot? By default \code{FALSE}}

\item{joint_only}{Only return the joint density plot.}

\item{axis_titles}{A string specifying how axis titltes should be treated.
\code{'keep'} will retain all axis titles in individual plots. \code{'collect'} will
remove duplicated titles in one direction and merge titles in the opposite
direction. \code{'collect_x'} and \code{'collect_y'} control this for x-axis titles
and y-axis titles respectively.}

\item{axes}{A string specifying how axes should be treated. \code{'keep'} will
retain all axes in individual plots. \code{'collect'} will remove duplicated
axes when placed in the same run of rows or columns of the layout.
\code{'collect_x'} and \code{'collect_y'} will remove duplicated x-axes in the columns
or duplicated y-axes in the rows respectively.}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not.  If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork (see example).}

\item{...}{
  Arguments passed on to \code{\link[Nebulosa:plot_density]{Nebulosa::plot_density}}
  \describe{
    \item{\code{object}}{Seurat or SingleCellExperiment object}
    \item{\code{slot}}{Type of data: \code{counts} or\code{data} for Seurat objects and
 \code{counts},
\code{logcounts}, or \code{normcounts} for SingleCellExperiment objects}
    \item{\code{reduction}}{Name of the reduction to visualize. If not provided, last
computed reduction is visualized}
    \item{\code{dims}}{Vector of length 2 specifying the dimensions to be plotted. By
default, the first two dimensions are considered.}
    \item{\code{adjust}}{Numeric value to adjust to bandwidth. Default: 1. Not available
for \code{ks} method}
    \item{\code{size}}{Size of the geom to be plotted (e.g. point size)}
    \item{\code{shape}}{Shape of the geom to be plotted}
    \item{\code{combine}}{Create a single plot? If \code{FALSE}, a list with ggplot
objects is returned}
    \item{\code{pal}}{String specifying the viridis color palette to use.}
    \item{\code{raster}}{Rasterise plot}
  }}
}
\description{
Plot feature density.
}
\examples{
obj <- get_HPO()
out <- plot_feature_density(obj, features =  c("C5","C7","C8A","C8B"))
}
