% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highlights.R
\name{plot_highlights}
\alias{plot_highlights}
\title{Plot highlights}
\usage{
plot_highlights(
  obj,
  name_col,
  queries,
  joint = TRUE,
  palettes = list("pals::brewer.blues", "pals::brewer.reds", "pals::brewer.purples",
    "pals::brewer.greens", "pals::brewer.oranges"),
  ignore.case = TRUE,
  ncol = 1,
  axis_titles = "collect",
  axes = "collect",
  guides = "collect",
  wrap_width = 20,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{axis_titles}{A string specifying how axis titltes should be treated.
\code{'keep'} will retain all axis titles in individual plots. \code{'collect'} will
remove duplicated titles in one direction and merge titles in the opposite
direction. \code{'collect_x'} and \code{'collect_y'} control this for x-axis titles
and y-axis titles respectively.}

\item{axes}{A string specifying how axes should be treated. \code{'keep'} will
retain all axes in individual plots. \code{'collect'} will remove duplicated
axes when placed in the same run of rows or columns of the layout.
\code{'collect_x'} and \code{'collect_y'} will remove duplicated x-axes in the columns
or duplicated y-axes in the rows respectively.}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not.  If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork (see example).}

\item{...}{
  Arguments passed on to \code{\link[Seurat:LabelPoints]{Seurat::LabelPoints}}
  \describe{
    \item{\code{plot}}{A ggplot2 plot with a GeomPoint layer}
    \item{\code{points}}{A vector of points to label; if \code{NULL}, will use all points in the plot}
    \item{\code{labels}}{A vector of labels for the points; if \code{NULL}, will use
rownames of the data provided to the plot at the points selected}
    \item{\code{repel}}{Use \code{geom_text_repel} to create a nicely-repelled labels; this
is slow when a lot of points are being plotted. If using \code{repel}, set \code{xnudge}
and \code{ynudge} to 0}
    \item{\code{xnudge,ynudge}}{Amount to nudge X and Y coordinates of labels by}
  }}
}
\description{
Plot highlighted subsets of the Seurat object.
}
\examples{
obj <- get_HPO()
out <- plot_highlights(obj,
                       name_col="HPO_label",
                       queries=c("Alzheimer","Parkinson"))
}
