% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ontological_similarity.R
\name{plot_ontological_similarity}
\alias{plot_ontological_similarity}
\title{Plot ontological similarity}
\usage{
plot_ontological_similarity(
  obj,
  id_col,
  ont,
  run_ontological_similarity_out = NULL,
  ancestors = unique(ont@elementMetadata$ancestor),
  group_var = "seurat_clusters",
  top_n = 9,
  min.cutoff = "q90",
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot}}
  \describe{
    \item{\code{object}}{Seurat object}
    \item{\code{features}}{Vector of features to plot. Features can come from:
\itemize{
  \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
  \item A column name from meta.data (e.g. mitochondrial percentage -
    "percent.mito")
  \item A column name from a \code{DimReduc} object corresponding to the
    cell embedding values (e.g. the PC 1 scores - "PC_1")
}}
    \item{\code{dims}}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}
    \item{\code{cells}}{Vector of cells to plot (default is all cells)}
    \item{\code{cols}}{The two colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high. Also accepts a Brewer
color scale or vector of colors. Note: this will bin the data into number of colors provided.
When blend is \code{TRUE}, takes anywhere from 1-3 colors:
\describe{
  \item{1 color:}{Treated as color for double-negatives, will use default colors 2 and 3 for per-feature expression}
  \item{2 colors:}{Treated as colors for per-feature expression, will use default color 1 for double-negatives}
  \item{3+ colors:}{First color used for double-negatives, colors 2 and 3 used for per-feature expression, all others ignored}
}}
    \item{\code{pt.size}}{Adjust point size for plotting}
    \item{\code{alpha}}{Alpha value for plotting (default is 1)}
    \item{\code{order}}{Boolean determining whether to plot cells in order of expression. Can be useful if
cells expressing given feature are getting buried.}
    \item{\code{min.cutoff,max.cutoff}}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}
    \item{\code{reduction}}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}
    \item{\code{split.by}}{A factor in object metadata to split the plot by, pass 'ident'
to split by cell identity'}
    \item{\code{keep.scale}}{How to handle the color scale across multiple plots. Options are:
\itemize{
  \item \dQuote{feature} (default; by row/feature scaling): The plots for
    each individual feature are scaled to the maximum expression of the
    feature across the conditions provided to \code{split.by}
  \item \dQuote{all} (universal scaling): The plots for all features and
    conditions are scaled to the maximum expression value for the feature
    with the highest overall expression
  \item \code{all} (no scaling): Each individual plot is scaled to the
    maximum expression value of the feature in the condition provided to
    \code{split.by}. Be aware setting \code{NULL} will result in color
    scales that are not comparable between plots
}}
    \item{\code{shape.by}}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells.  Only applicable if \code{raster = FALSE}.}
    \item{\code{slot}}{Which slot to pull expression data from?}
    \item{\code{blend}}{Scale and blend expression values to visualize coexpression of two features}
    \item{\code{blend.threshold}}{The color cutoff from weak signal to strong signal; ranges from 0 to 1.}
    \item{\code{label}}{Whether to label the clusters}
    \item{\code{label.size}}{Sets size of labels}
    \item{\code{label.color}}{Sets the color of the label text}
    \item{\code{repel}}{Repel labels}
    \item{\code{ncol}}{Number of columns to combine multiple feature plots to, ignored if \code{split.by} is not \code{NULL}}
    \item{\code{coord.fixed}}{Plot cartesian coordinates with fixed aspect ratio}
    \item{\code{by.col}}{If splitting by a factor, plot the splits per column with the features as rows; ignored if \code{blend = TRUE}}
    \item{\code{sort.cell}}{Redundant with \code{order}. This argument is being
deprecated. Please use \code{order} instead.}
    \item{\code{interactive}}{Launch an interactive \code{\link[Seurat:IFeaturePlot]{FeaturePlot}}}
    \item{\code{combine}}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}
ggplot object. If \code{FALSE}, return a list of ggplot objects}
    \item{\code{raster}}{Convert points to raster format, default is \code{NULL} which
automatically rasterizes if plotting more than 100,000 cells}
    \item{\code{raster.dpi}}{Pixel resolution for rasterized plots, passed to geom_scattermore().
Default is c(512, 512).}
  }}
}
\description{
Plot ontological similarity of cells to a set of terms in an ontology.
}
