% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_preservation_histo.R
\name{plot_preservation_histo}
\alias{plot_preservation_histo}
\title{Plot preservation histogram}
\usage{
plot_preservation_histo(
  consist_dt,
  lvl = "hd",
  key = c(hd = "High-dimensional", md = "Mid-dimensional", ld = "Low-dimensional"),
  title = paste(key[lvl], "space"),
  subtitle = paste0("(", consist_dt[[paste0("dim_", lvl)]][1], " dimensions)"),
  xlim = c(-1, 1)
)
}
\description{
Plot the correlation between equivalent terms in 
high-, mid-, and low-dimensional space.
}
