% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{Plot reduction}
\usage{
plot_reduction(
  obj,
  reduction = NULL,
  metadata = NULL,
  fix_rownames = FALSE,
  color_var = NULL,
  label_var = NULL,
  size_var = NULL,
  labels = TRUE,
  max.overlaps = 30,
  show_plot = TRUE,
  point_alpha = 0.6
)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{Seurat}}
}}

\item{reduction}{Which reduction to extract from.
Only used if \code{obj} is of class \code{Seurat}.}
}
\value{
ggplot object
}
\description{
Plot dimensionality reduction results
from \code{run_<method>} or a Seurat object.
}
