% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top.R
\name{plot_top}
\alias{plot_top}
\title{Plot the top factors}
\usage{
plot_top(
  top_data,
  factors = NULL,
  x = "label_phe",
  y = "loading",
  fill = "factor",
  title = NULL,
  invert_vars = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{top_data}{Data with the top phenotypes or features.}

\item{factors}{Which factors to plot.}

\item{x}{x-axis variable.}

\item{y}{y-axis variable.}

\item{fill}{fill variable.}

\item{title}{Plot title.}

\item{invert_vars}{Switch the axes of factors and loadings in plot.}

\item{show_plot}{Whether to print the plot or simply return it.}
}
\value{
ggplot object
}
\description{
Plot the top factors for a given set of phenotypes, 
OR plot the top features for a given set of factors.
}
\keyword{internal}
