% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_celltypes.R
\name{plot_top_celltypes}
\alias{plot_top_celltypes}
\title{Plot top cell types}
\usage{
plot_top_celltypes(
  obj,
  types = c("radial", "pie", "reduction"),
  reduction = "umap",
  prefix = "^q[.]",
  cluster_vars = c("seurat_clusters"),
  color_col = cluster_vars[1],
  label_col = cluster_vars[1],
  enriched_proportion_threshold = 0.2,
  label_alpha = 0.9,
  label_size = 3,
  label_color = "white"
)
}
\description{
Plot the top cell types per cluster.
}
\examples{
obj <- get_HPO()
out <- plot_top_celltypes(obj)
}
