% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_cor.R
\name{plot_trait_cor}
\alias{plot_trait_cor}
\title{Plot top trait-trait correlations}
\usage{
plot_trait_cor(knn, top_n = 10, non_self = TRUE, show_plot = TRUE)
}
\arguments{
\item{knn}{A melted similarity matrix or K-nearest neighbor graph.}

\item{top_n}{The number of top correlations to plot.}

\item{non_self}{Remove trait2 that are present in trait1.}

\item{show_plot}{Whethe to print the plot.}
}
\value{
\code{ggplot} object.
}
\description{
Takes the output of \link[phenomix]{find_neighbors} as input to \code{knn}.
}
