% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variancePartition.R
\name{plot_variancePartition}
\alias{plot_variancePartition}
\title{Plot \pkg{variancePartition}}
\usage{
plot_variancePartition(
  varPart,
  plot_PercentBars = TRUE,
  plot_VarPart = TRUE,
  genes = NULL,
  n_genes = 10,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{varPart}{\link[variancePartition]{fitExtractVarPartModel}
output.}

\item{plot_PercentBars}{Add  \link[variancePartition]{plotPercentBars}.}

\item{plot_VarPart}{Add  \link[variancePartition]{plotPercentBars}.}

\item{genes}{Genes to plot in \link[variancePartition]{plotVarPart}.}

\item{n_genes}{Number of genes to select when \code{genes=NULL}.}

\item{...}{Additional arguments passed to
\link[patchwork]{wrap_plots}.}
}
\description{
Plot \link[variancePartition]{fitExtractVarPartModel}
output.
}
