% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.phenomix}
\alias{predict.phenomix}
\title{Get predicted values from model object, copying glmmTMB 'fast' implementation}
\usage{
\method{predict}{phenomix}(object, se.fit = FALSE, ...)
}
\arguments{
\item{se.fit}{Boolean, whether to produce SEs defaults to FALSE}

\item{...}{Extra parameters}

\item{fit}{The fitted phenomix model}
}
\description{
Get predicted values from model object, copying glmmTMB 'fast' implementation
}
