% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_agg_funcs.R
\name{prepare_agg_funcs}
\alias{prepare_agg_funcs}
\title{prepare_agg_funcs}
\usage{
prepare_agg_funcs(
  merged_hits,
  extra_cols = c("ABC.Score", "CellType"),
  verbose = TRUE
)
}
\description{
prepare_agg_funcs
}
\keyword{internal}
