% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hpo.R
\name{prepare_hpo}
\alias{prepare_hpo}
\title{Prepare Human Phenotype Ontology}
\usage{
prepare_hpo(
  dt_genes = HPOExplorer::load_phenotype_to_genes(1),
  dt_annot = HPOExplorer::load_phenotype_to_genes(3),
  ctd = NULL,
  ct_results = MultiEWCE::load_example_results("Descartes_All_Results_extras.rds"),
  id_types = c("disease_id", "hpo_id"),
  min_quantile = 35,
  min_genes = NULL,
  value.var = "evidence_score_mean",
  vars.to.regress = "n_genes",
  run_nlp = FALSE,
  run_impute = FALSE,
  workers = 1,
  seed = 2023,
  verbose = TRUE
)
}
\description{
Conver the Human Phenotype Ontology (HPO) to a Seurat object with
extensive metadata.
}
\examples{
obj <- prepare_hpo(id_types="hpo_id")
}
