% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_primekg.R
\name{prepare_primekg}
\alias{prepare_primekg}
\title{Prepare PrimeKG}
\usage{
prepare_primekg(
  types = c("gene_concept", "gene_concept_imputed", "gene_gene", "seurat", "scNLP"),
  fill = 0,
  impute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{types}{Character vector of types of data to return.}

\item{fill}{Value with which to fill missing cells. If \code{fun.aggregate} is present, takes the value by applying the function on a 0-length vector.}

\item{impute}{Impute missing values in \code{gene_concept} matrix 
using the \code{gene_gene} matrix. If \code{TRUE}, the imputed dataset 
will be used as a basis of the \code{seurat} object (if selected).}

\item{verbose}{Print messages.}
}
\value{
\code{Seurat} object
}
\description{
Prepare the \href{https://github.com/mims-harvard/PrimeKG}{PrimeKG} dataset
 as a \pkg{Seurat} object.
}
