% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qdgnorm}
\alias{qdgnorm}
\title{Calculate the quantile of a double general normal distribution}
\usage{
qdgnorm(p, mu, sigma1, sigma2, beta_ratio_1, beta_ratio_2, beta_1, beta_2)
}
\arguments{
\item{p}{The value to evaluate the quantile function at}

\item{mu}{The mean of the distribution}

\item{sigma1}{The sigma of the left half of the distribution}

\item{sigma2}{The sigma of the right half of the distribution}

\item{beta_ratio_1}{The beta ratio the left half of the distribution}

\item{beta_ratio_2}{The beta ratio of the right half of the distribution}

\item{beta_1}{The beta parameter of the left half of the distribution}

\item{beta_2}{The beta parameter of the right half of the distribution}
}
\description{
Calculate the quantile of a double general normal distribution
}
