% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qdnorm}
\alias{qdnorm}
\title{Calculate the quantile of a double normal distribution}
\usage{
qdnorm(p, mu, sigma1, sigma2)
}
\arguments{
\item{p}{The value to evaluate the quantile function at}

\item{mu}{The mean of the distribution}

\item{sigma1}{The sigma of the left half of the distribution}

\item{sigma2}{The sigma of the right half of the distribution}
}
\description{
Calculate the quantile of a double normal distribution
}
