% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomix_rtmb.R
\name{qthill}
\alias{qthill}
\title{RTMB Objective Function for phenomix}
\usage{
qthill(quantile, v, mean, sigma)
}
\arguments{
\item{pars}{List of parameters}

\item{data}{List of data}
}
\value{
Negative log-likelihood
}
\description{
This function replaces the TMB C++ template with an R-based RTMB function
}
\keyword{internal}
