% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ranef.phenomix}
\alias{ranef.phenomix}
\title{Get random effects parameters from model object, copying glmmTMB 'fast' implementation}
\usage{
\method{ranef}{phenomix}(object, ...)
}
\arguments{
\item{object}{The fitted phenomix model}

\item{...}{Additional arguments}
}
\description{
Get random effects parameters from model object, copying glmmTMB 'fast' implementation
}
