% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_matrix.R
\name{reconstruct_matrix}
\alias{reconstruct_matrix}
\title{Reconstruct original matrix from DeGAs components}
\usage{
reconstruct_matrix(
  u_name = "contribution_phe",
  v_name = "contribution_var",
  d_name = NULL,
  DEGAS = NULL,
  annot = NULL,
  gene_level = FALSE,
  remove_nongenes = TRUE,
  filter_quantiles = FALSE,
  norm = TRUE,
  norm_scale = c(1, 100),
  replace_underscores = TRUE
)
}
\description{
Reconstruct original matrix from DeGAs components
}
\keyword{internal}
