% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cor.R
\name{run_cor}
\alias{run_cor}
\title{Run correlation}
\usage{
run_cor(X, t = FALSE)
}
\arguments{
\item{X}{Matrix to compute correlations}

\item{t}{Transpose the matrix.}
}
\description{
Run pairwise correlations on a matrix.
}
