% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dag_enrich.R
\name{run_dag_enrich}
\alias{run_dag_enrich}
\title{Run enrichment of terms in a DAG ontology}
\usage{
run_dag_enrich(
  obj,
  ont,
  reduction = NULL,
  cluster_col = NULL,
  min_hits = 3,
  min_offspring = 100,
  trans_fun = NULL,
  replace_char = list(. = ":", `_` = ":"),
  q_threshold = 0.05,
  top_n = 100
)
}
\description{
Run enrichment of terms in a DAG ontology.
}
\examples{
ont <- KGExplorer::get_ontology("hp")
obj <- get_HPO()
res <- run_dag_enrich(obj, ont, reduction="pca")
}
