% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gprofiler.R
\name{run_gprofiler}
\alias{run_gprofiler}
\title{Run gprofiler}
\usage{
run_gprofiler(
  obj,
  n_features = 3,
  factors = seq(1, 4),
  show_plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{\pkg{Seurat} object or dimensionality reduction object.}

\item{n_features}{Number of top features per factor to select.}

\item{factors}{Which factors to run gprofiler on.}

\item{show_plot}{Whether to print the plot or simply return it.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[gprofiler2]{gost}.}
}
\description{
Run \link[gprofiler2]{gost} to perform
 Gene Set  Enrichment Analysis (GSEA).
}
