% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_integration.R
\name{run_integration}
\alias{run_integration}
\title{Run integration}
\usage{
run_integration(
  obj,
  split.by,
  assay = Seurat::DefaultAssay(obj),
  nfeatures = nrow(obj),
  vars.to.regress = NULL,
  dims = seq(100),
  save_path = tempfile(fileext = ".rds"),
  pipeline = c("seuratv5", "seuratv4"),
  method = Seurat::CCAIntegration,
  k.weight = 100,
  orig.reduction = "pca",
  new.reduction = "cca",
  cluster_reduction = "umap",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{split.by}{Attribute for splitting. Default is "ident". Currently
only supported for class-level (i.e. non-quantitative) attributes.}

\item{assay}{Name of assay for integration}

\item{nfeatures}{Number of features to select as top variable features;
only used when \code{selection.method} is set to \code{'dispersion'} or
\code{'vst'}}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in
RegressOut). For example, nUMI, or percent.mito.}

\item{dims}{Dimensions of dimensional reduction to use for integration}

\item{method}{Integration method function}

\item{k.weight}{Number of neighbors to consider when weighting anchors}

\item{orig.reduction}{Name of dimensional reduction for correction}

\item{new.reduction}{Name of new integrated dimensional reduction}

\item{cluster_reduction}{Recompute neighbors graph based on UMAP
to get clusters that best reflect UMAP space.
For this same reason, only cluster in two dimensions,
because this is the view we most often use.
That said, this may reduce the generalisability of these clusters/graph.}

\item{verbose}{Displays a progress bar for scaling procedure}

\item{...}{
  Arguments passed on to \code{\link[Seurat:IntegrateLayers]{Seurat::IntegrateLayers}}
  \describe{
    \item{\code{object}}{A \code{\link[SeuratObject]{Seurat}} object}
    \item{\code{features}}{A vector of features to use for integration}
    \item{\code{layers}}{Names of normalized layers in \code{assay}}
    \item{\code{scale.layer}}{Name(s) of scaled layer(s) in \code{assay}}
  }}
}
\description{
Run integration pipeline on a \link{Seurat} object.
}
\examples{
obj <- scKirby::example_obj("seurat")
obj2 <- run_integration(obj, split.by="groups")
}
