% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lda.R
\name{run_lda}
\alias{run_lda}
\title{Run LDA}
\source{
\url{https://towardsdatascience.com/linear-discriminant-analysis-lda-101-using-r-6a97217a55a6}
}
\usage{
run_lda(
  mat,
  grouping,
  transpose = TRUE,
  center = TRUE,
  scale. = FALSE,
  rank. = NULL,
  ...
)
}
\arguments{
\item{mat}{Matrix to run PCA on.}

\item{grouping}{
    (required if no formula principal argument is given.)
    a factor specifying the class for each observation.
  }

\item{transpose}{Whether to transpose the matrix first.}

\item{...}{Additional parameters passed to \link[MASS]{lda}.}
}
\description{
Run Linear Discriminant Analysis (LDA).
\emph{WARNING:} Takes a very long time to run on medium to large datasets.
}
\details{
Uses \link[stats]{prcomp}.
}
\examples{
mat <- as.matrix(iris[, seq(1, 4)])
grouping <- iris$Species
}
