% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mofa2.R
\name{run_mofa2}
\alias{run_mofa2}
\title{Run MOFA2}
\source{
\href{https://biofam.github.io/MOFA2/}{MOFA2 site}
}
\usage{
run_mofa2(
  obj,
  groups = NULL,
  obs = NULL,
  assay = NULL,
  layer = NULL,
  features = "variable_features",
  obs_idcol = "sample",
  transpose = FALSE,
  maxiter = 1000,
  num_factors = NULL,
  data_options = NULL,
  model_options = NULL,
  training_options = NULL,
  stochastic_options = NULL,
  mefisto_options = NULL,
  outfile = tempfile(fileext = ".mofa_model.hdf5"),
  reductions = c("umap"),
  use_basilisk = TRUE,
  seed = 2020,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{Named list of input data matrices, or a \pkg{Seurat} object.}

\item{groups}{group information, only relevant when using the multi-group framework.}

\item{obs}{Metadata associated with \code{obj}. 
Will be extracted automatically if \code{obj} is a \pkg{Seurat} object.}

\item{assay}{Assay to use in \code{obj}.}

\item{layer}{Layer to use in \code{obj}.}

\item{features}{Features to use in reduction. Can be:
\itemize{
\item{\<vector\> : }{A character vector of feature names.}
\item{"variable_features" : }{Will trigger the automatic extraction of
 precalculated variable features
  (via \link[scKirby]{get_variable_features}).}
\item{\code{NULL} : }{Uses all features.}
}}

\item{obs_idcol}{The name of the column to add to \code{obs} 
after fixing IDs.}

\item{transpose}{First transpose each matrix in \code{mat_list}.}

\item{maxiter}{Maximum number of training iterations (\emph{DEFAULT} = 1000).}

\item{data_options}{list of data_options (see \code{\link[MOFA2]{get_default_data_options}} details). 
If NULL, default options are used.}

\item{model_options}{list of model options (see \code{\link[MOFA2]{get_default_model_options}} for details). 
If NULL, default options are used.}

\item{training_options}{list of training options (see \code{\link[MOFA2]{get_default_training_options}} for details). 
If NULL, default options are used.}

\item{stochastic_options}{list of options for stochastic variational inference (see \code{\link[MOFA2]{get_default_stochastic_options}} for details). 
If NULL, default options are used.}

\item{mefisto_options}{list of options for mefisto (see \code{\link[MOFA2]{get_default_mefisto_options}} for details). 
If NULL, default options are used.}

\item{outfile}{output file for the model (.hdf5 format). If \code{NULL}, a temporary file is created.}

\item{reductions}{In addition to returning MOFA factors,
further reduce the data by "umap" and/or "tsne" to the MOFA factors.}

\item{use_basilisk}{use \code{basilisk} to automatically install a conda environment with mofapy2 and all dependencies? 
If \code{FALSE} (default), you should specify the right python binary when loading R with \code{reticulate::use_python(..., force=TRUE)}
or the right conda environment with \code{reticulate::use_condaenv(..., force=TRUE)}.}

\item{seed}{Seed passed to \link[base]{set.seed}
for reproducibility between runs.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[MOFA2:run_mofa]{MOFA2::run_mofa}}
  \describe{
    \item{\code{object}}{an untrained \code{\link[MOFA2]{MOFA}} object}
    \item{\code{save_data}}{logical indicating whether to save the training data in the hdf5 file. 
This is useful for some downstream analysis (mainly functions with the prefix \code{plot_data}), but it can take a lot of disk space.}
  }}
}
\description{
Run Multi-Omics Factor Analysis v2 (MOFA2).
}
\details{
Uses \link[MOFA2]{run_mofa} to reduce a single dataset,
or multiple views of the same dataset, into a
set number of factors.
}
\examples{
obj <- get_HPO()[seq(100),]
res <- run_mofa2(obj, maxiter=10) 
}
