% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_preseveration.R
\name{run_preservation}
\alias{run_preservation}
\title{Run preservation}
\usage{
run_preservation(
  obj,
  reduction = list(md = "pca", ld = "umap"),
  id_types = c("hp_id", "mondo_id", "efo_id"),
  k.param = 30,
  dims = list(md = 1:10, ld = 1:2),
  cor_fun = run_cor,
  max_ids = 10000,
  save_path = NULL,
  force_new = FALSE
)
}
\arguments{
\item{obj}{Seurat object.}

\item{reduction}{List of reductions to use for preservation.}

\item{id_types}{List of id types to use for preservation.}

\item{k.param}{Number of nearest neighbours to use for preservation.}

\item{dims}{List of dimensions to use for preservation.}

\item{cor_fun}{Function to use for correlation.}

\item{max_ids}{Maximum number of ids to use for preservation.}
}
\value{
data.table with preservation results.
}
\description{
Assess the preservation of high-dimensional structure in 
mid-dimensional and low-dimensional space.
}
\examples{
obj <- get_HPO()
obj$id[1:100] <- obj$id[1] 
res <- run_preservation(obj, id_types="id")
}
