% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pseudotime_subtypes.R
\name{run_pseudotime_subtypes}
\alias{run_pseudotime_subtypes}
\title{Run pseudotime: subtypes pipeline}
\usage{
run_pseudotime_subtypes(
  obj,
  query,
  disease_ids = NULL,
  disease_ids_sub = NULL,
  title = query,
  id_col = "hp_id",
  name_col = "name",
  p2g = HPOExplorer::add_disease(HPOExplorer::load_phenotype_to_genes()),
  reduction_ld = "umap",
  n_genes = 50,
  max_disease_ids = 3,
  min_gene_sum = 0,
  future.globals.maxSize = (60 * 1000) * 1024^2,
  widths = c(1, 0.25),
  show_plot = TRUE,
  force_new = FALSE,
  plot_path = tempfile(fileext = ".png"),
  data_path = tempfile(fileext = ".csv.gz"),
  width = 10,
  height = 16
)
}
\description{
Run pseudotime trajectory analysis on a disease and several of its subtypes.
}
\examples{
obj <- get_HPO()
id_col <- "id"
name_col <- "hpo_name"
}
