% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sparsesvd.R
\name{run_sparsesvd}
\alias{run_sparsesvd}
\title{Run sparse SVD}
\usage{
run_sparsesvd(
  mat,
  transpose = TRUE,
  add_names = TRUE,
  rank = 0L,
  tol = 1e-15,
  kappa = 1e-06
)
}
\arguments{
\item{mat}{Matrix to run sparseSVD on.}

\item{transpose}{Whether to transpose the matrix first.}

\item{add_names}{Add colnames and rownames to embeddings and loadings.}

\item{rank}{an integer specifying the desired number of singular components, i.e. the rank of the truncated SVD.
    Specify 0 to return all singular values of magnitude larger than \code{tol} (default).
  }

\item{tol}{exclude singular values whose magnitude is smaller than \code{tol}}

\item{kappa}{accuracy parameter \eqn{\kappa} of the SVD algorithm (with SVDLIBC default)}
}
\description{
Run sparse Singular Value Decomposition (sparseSVD).
}
\details{
Uses \link[sparsesvd]{sparsesvd}.
}
