% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_variancePartition.R
\name{run_variancePartition}
\alias{run_variancePartition}
\title{Run \pkg{variancePartition}}
\usage{
run_variancePartition(
  obj,
  metadata = NULL,
  workers = NULL,
  formula = NULL,
  is_opengwas = FALSE,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{Matrix or \pkg{Seurat} object to run
\link[variancePartition]{fitExtractVarPartModel} on.}

\item{metadata}{\code{data.frame} containing metadata
to use in \code{formula}. If \code{obj} is a \pkg{Seurat} object,
this can be left \code{NULL} and metadata
will be extracted automatically.}

\item{workers}{
    \code{integer(1)} Number of workers. Defaults to the maximum of 1 or
    the number of cores determined by \code{detectCores} minus 2 unless
    environment variables \code{R_PARALLELLY_AVAILABLECORES_FALLBACK} or
    \code{BIOCPARALLEL_WORKER_NUMBER} are set otherwise.
  }

\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used as a response. e.g.: \code{~ a + b + (1|c)}}

\item{is_opengwas}{Whether the data comes from
 \href{https://gwas.mrcieu.ac.uk/}{OpenGWAS}.
If \code{TRUE}, a predefined formula and metadata
processing procedure will be used.}

\item{show_plot}{Print plot.}

\item{...}{Additional arguments passed to
\link[variancePartition]{fitExtractVarPartModel}.}
}
\value{
\code{varPart} results.
}
\description{
Run \link[variancePartition]{fitExtractVarPartModel}
to evaluate the effect of each metadata variable on your data.
}
