% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save ggplot}
\usage{
save_plot(save_path, plot, width = NA, height = NA, dpi = 300)
}
\arguments{
\item{save_path}{Path to save plot to.}

\item{plot}{ggplot object}

\item{width}{Plot width}

\item{height}{Plot height.}

\item{dpi}{Plot resolution.}
}
\description{
Save ggplot
}
\keyword{internal}
