% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_ontological_similarity.R
\name{seurat_to_ontological_similarity}
\alias{seurat_to_ontological_similarity}
\title{Seurat to ontological similarity}
\usage{
seurat_to_ontological_similarity(
  obj,
  id_col,
  ont,
  ancestors,
  group_var = NULL,
  as_sparse = TRUE,
  return_assay = TRUE,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[simona:term_sim]{simona::term_sim}}
  \describe{
    \item{\code{dag}}{An \code{ontology_DAG} object.}
    \item{\code{terms}}{A vector of term names.}
    \item{\code{method}}{A term similarity method. All available methods are in \code{\link[simona:all_term_sim_methods]{all_term_sim_methods()}}.}
    \item{\code{control}}{A list of parameters passing to individual methods. See the subsections.}
    \item{\code{verbose}}{Whether to print messages.}
  }}
}
\description{
Takes a Seurat object and computes the ontological similarity of 
each sample (column) given some ontology that it can be mapped onto.
}
