% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nightlight.R
\name{theme_nightlight}
\alias{theme_nightlight}
\title{Theme nightlight}
\usage{
theme_nightlight(text = "white", background = "black", color = "")
}
\arguments{
\item{text}{Text color.}

\item{background}{Background color.}
}
\value{
ggplot2 theme.
}
\description{
\link{ggplot2} theme with dark background and white text.
}
