% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_geneids_txdb.R
\name{translate_geneids_txdb}
\alias{translate_geneids_txdb}
\title{Translate GENEId to SYMBOL}
\usage{
translate_geneids_txdb(
  gene_hits,
  gene_var = "GENEID",
  drop_na = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gene_hits}{Output from \code{aggregate_sumstats}.}

\item{gene_var}{Column name with gene IDs.}

\item{drop_na}{Drop genes without a corresponding \code{SYMBOL}}

\item{verbose}{Print messages.}
}
\description{
Translate \code{GENEID} from 
\link[TxDb.Hsapiens.UCSC.hg19.knownGene]{TxDb.Hsapiens.UCSC.hg19.knownGene}
or
\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}
to gene \code{SYMBOL}.
}
\keyword{internal}
