% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cores.r
\name{assign_cores}
\alias{assign_cores}
\title{Assign cores}
\usage{
assign_cores(worker_cores = 0.9, verbose = TRUE)
}
\description{
Automatically infer the number of available cores and
 allocate the majority of them to parallel processing.
 Reserves a small proportion of cores to avoid slowing or crashing.
}
\keyword{internal}
