% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Create data file for fitting time varying run timing distributions with TMB}
\usage{
create_data(
  data,
  min_number = 0,
  variable = "number",
  time = "year",
  date = "doy",
  asymmetric_model = TRUE,
  mu = ~1,
  sigma = ~1,
  covar_data = NULL,
  est_sigma_re = TRUE,
  est_mu_re = TRUE,
  tail_model = "gaussian",
  family = "gaussian",
  max_theta = 10,
  share_shape = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{min_number}{A minimum threshold to use, defaults to 0}

\item{variable}{A character string of the name of the variable in 'data' that contains the response (e.g. counts)}

\item{time}{A character string of the name of the variable in 'data' that contains the time variable (e.g. year)}

\item{date}{A character string of the name of the variable in 'data' that contains the response (e.g. day of year). The actual
#' column should contain a numeric response -- for example, the result from using lubridate::yday(x)}

\item{asymmetric_model}{Boolean, whether or not to let model be asymmetric (e.g. run timing before peak has a
different shape than run timing after peak)}

\item{mu}{An optional formula allowing the mean to be a function of covariates. Random effects are not included in the formula
but specified with the \code{est_mu_re} argument}

\item{sigma}{An optional formula allowing the standard deviation to be a function of covariates. For asymmetric models,
each side of the distribution is allowed a different set of covariates. Random effects are not included in the formula
but specified with the \code{est_sigma_re} argument}

\item{covar_data}{a data frame containing covariates specific to each time step. These are used in the formulas \code{mu} and \code{sigma}}

\item{est_sigma_re}{Whether to estimate random effects by year in sigma parameter controlling tail of distribution. Defaults to TRUE}

\item{est_mu_re}{Whether to estimate random effects by year in mu parameter controlling location of distribution. Defaults to TRUE}

\item{tail_model}{Whether to fit Gaussian ("gaussian" = default) or Student-t ("student_t") or generalized normal ("gnorm"). Defaults to "gaussian"}

\item{family}{Response for observation model, options are "gaussian", "poisson", "negbin"}

\item{max_theta}{Maximum value of log(pred) when \code{limits=TRUE}. Defaults to 10}

\item{share_shape}{Boolean argument for whether asymmetric student-t and generalized normal distributions should share the shape parameter (nu for the student-t;
beta for the generalized normal). Defaults to TRUE}
}
\description{
Does minimal processing of data to use as argument to fitting function
}
\examples{
data(fishdist)
datalist <- create_data(fishdist,
  min_number = 0, variable = "number", time = "year",
  date = "doy", asymmetric_model = TRUE, family = "gaussian"
)
}
