% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cor.R
\name{extract_cor}
\alias{extract_cor}
\title{Extract correlation matrix}
\usage{
extract_cor(
  obj,
  assay = NULL,
  slot = NULL,
  reduction = NULL,
  graph_name = NULL,
  method = "pearson",
  return_obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Either a \pkg{Seurat} object or a feature*trait matrix.}

\item{assay}{Assay to used to compute correlation graph 
if one does not already exist.}

\item{slot}{If \code{obj} is a \pkg{Seurat} object, which slot to extract
for correlations.}

\item{reduction}{If not \code{NULL}, extracts the embedding from the
specified reduction (e.g. "pca") and uses that to
compute trait-trait similarities instead.}

\item{graph_name}{Name of the graph to extract.}

\item{method}{Pairwise correlation method.}

\item{return_obj}{Whether to return the \pkg{Seurat} object with a new
\code{obj@graphs} slot, or to simply return the sparse correlation matrix.}

\item{verbose}{Print messages.}
}
\value{
Trait-trait correlation matrix.
}
\description{
Extract correlation matrix stored as a graph.
}
\examples{
hpo <- phenomix::get_HPO()
cmat <- extract_cor(obj = hpo,
                    reduction = "pca")
}
