% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishdist}
\alias{fishdist}
\title{Example simulate data for fish distributions from multiple years}
\format{
A data frame containing simulated data.
}
\usage{
fishdist
}
\description{
Example simulate data for fish distributions from multiple years
}
\keyword{internal}
