% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fitting function to be called by user}
\usage{
fit(
  data_list,
  silent = FALSE,
  inits = NULL,
  control = list(eval.max = 2000, iter.max = 1000, rel.tol = 1e-10),
  limits = NULL,
  fit_model = TRUE
)
}
\arguments{
\item{data_list}{A list of data, as output from create_data}

\item{silent}{Boolean passed to RTMB::MakeADFun, whether to be verbose or not (defaults to FALSE)}

\item{inits}{Optional named list of parameters for starting values, defaults to NULL}

\item{control}{Optional control list for stats::nlminb. For arguments see ?nlminb. Defaults to eval.max=2000, iter.max=1000, rel.tol=1e-10. For final model runs, the rel.tol should be even smaller}

\item{limits}{Whether to include limits for stats::nlminb. Can be a list of (lower, upper), or TRUE to use suggested hardcoded limits. Defaults to NULL,
where no limits used.}

\item{fit_model}{Whether to fit the model. If not, returns a list including the data, parameters, and initial values. Defaults to TRUE}
}
\description{
This function creates a list of parameters, sets up RTMB object and attempts to
do fitting / estimation
}
\examples{
\dontrun{
data(fishdist)

# example of fitting fixed effects, no trends, no random effects
set.seed(1)
datalist <- create_data(fishdist[which(fishdist$year > 1970), ],
  asymmetric_model = FALSE,
  est_mu_re = FALSE, est_sigma_re = FALSE
)
fit <- fit(datalist)
}
}
