% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fitting function to be called by user}
\usage{
fit(data_list, silent = FALSE)
}
\arguments{
\item{data_list}{A list of data, as output from create_data}

\item{silent}{Boolean passed to TMB::MakeADFun, whether to be verbose or not (defaults to FALSE)}
}
\description{
This function creates a list of parameters, sets up TMB object and attempts to
do fitting / estimation
}
\examples{
data(fishdist)
datalist = create_data(fishdist)
}
