% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd.R
\name{get_ctd}
\alias{get_ctd}
\title{Get a CellTypeDataset}
\usage{
get_ctd(file = "BlueLake2018_FrontalCortexOnly.rds")
}
\arguments{
\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}
}
\value{
\pkg{CTD} object.
}
\description{
Get a CellTypeDataset object.
}
\examples{
ctd <- get_ctd()
}
