% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_factors.R
\name{get_top_factors}
\alias{get_top_factors}
\title{Get the top factors}
\usage{
get_top_factors(
  obj,
  metadata = NULL,
  reduction = NULL,
  term,
  search_col = "label_phe",
  n_quantiles = 10,
  select_quantiles = n_quantiles,
  plot_hist = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{\pkg{Seurat} object or dimensionality reduction object.}

\item{metadata}{Phenotype metadata. 
Not needed if \code{obj} is a \pkg{Seurat} object.}

\item{reduction}{Which reduction to extract from. 
Only used if \code{obj} is of class \code{Seurat}.}

\item{term}{Term with which to perform substring search of observations.}

\item{search_col}{Which column in the observation metadata to perform substring search.}

\item{n_quantiles}{How many quantiles to bin factor loadings into.}

\item{select_quantiles}{Which quantiles to return. Defaults to the top quantile only.}

\item{plot_hist}{Whether to plot the distribution of loadings.}

\item{verbose}{Print messages.}
}
\value{
data.table
}
\description{
Search a reduction for observations (e.g. traits) that match a given \code{term} (case-insensitive substring search).
Then get the factors with the highest mean loading for the matching features.
}
\examples{
data("DEGAS_seurat")
top_factors <- get_top_factors(obj=DEGAS_seurat,
                               term = "parkinson",
                               select_quantiles=8:10)
}
