% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_traits.R
\name{get_top_traits}
\alias{get_top_traits}
\title{Plot the top traits}
\usage{
get_top_traits(
  obj,
  obs = NULL,
  keys = NULL,
  n_traits = 3,
  factors = seq(10),
  invert_vars = FALSE,
  show_plot = TRUE,
  title = NULL,
  x = "trait",
  y = "loading",
  verbose = TRUE
)
}
\arguments{
\item{obj}{\pkg{Seurat} object or dimensionality reduction object.}

\item{obs}{Phenotype metadata.
Not needed if \code{obj} is a \pkg{Seurat} object.}

\item{keys}{The keys of reductions to extract from.}

\item{n_traits}{Number of top traits per reduction factor to select.}

\item{factors}{Which factors to plot.}

\item{invert_vars}{Switch the axes of factors and loadings in plot.}

\item{show_plot}{Whether to print the plot or simply return it.}

\item{title}{Plot title.}

\item{x}{x-axis variable.}

\item{y}{y-axis variable.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} of top traits.
}
\description{
Create a radar chart of the traits (phenotypes) with the
highest loadings for each reduction factor.
}
\examples{
obj <- get_HPO()
top_phenos <- get_top_traits(obj = obj)
}
