% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_lm.R
\name{iterate_lm}
\alias{iterate_lm}
\title{Iterate linear regression across all traits and celltypes}
\usage{
iterate_lm(
  xmat,
  ymat,
  correction_method = "BH",
  qvalue_thresh = 0.05,
  quantize = list(x = NULL, y = NULL),
  progressbar = TRUE,
  workers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{xmat}{gene x trait matrix.}

\item{ymat}{gene x celltype matrix.}

\item{correction_method}{Multiple-testing correction 
method to be passed to \code{stats::p.adjust}.}

\item{qvalue_thresh}{q-value threshold to use when report
significant results summary.}

\item{quantize}{A named list where the values of "x" and "y" indicaite the 
number of quantiles to bin the respective "xmat" and "ymat" datasets into.}

\item{progressbar}{logical(1) Enable progress bar 
(based on \code{plyr:::progress_text}).
 Enabling the progress bar changes the default value of tasks to 
 \code{.Machine$integer.max}, so that progress is reported for 
 each element of X.}

\item{workers}{Number (>1) or proportion (<1) of worker cores to use.}

\item{verbose}{Print messages.}
}
\description{
Iterate linear regression across all traits and celltypes
}
\examples{
### Phenotype dataset
obj <- get_HPO()
xmat <- scKirby::get_x(obj)[["RNA.data"]]
xmat <- xmat[,seq(10)] # Let's use just 10 traits as an example

### Celltypes Dataset
ctd <- get_ctd()
ymat <- ctd[[1]]$specificity
res_lm <- iterate_lm(xmat = xmat,
                     ymat = ymat, 
                     workers = 1)
}
