% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_matrix.R
\name{magma_matrix}
\alias{magma_matrix}
\title{Construct gene x trait matrix from MAGMA results}
\usage{
magma_matrix(
  magma_out,
  metric = c("ADJ_ZSTAT", "ZSTAT", "P"),
  drop_mhc = TRUE,
  save_path = NULL,
  fillna = FALSE,
  trans_fun = NULL,
  agg_fun = "mean",
  all = TRUE,
  as_matrix = TRUE,
  workers = 1,
  verbose = TRUE
)
}
\arguments{
\item{magma_out}{Paths to \emph{genes.out} files
produced by \link[MAGMA.Celltyping]{map_snps_to_genes}.}

\item{metric}{Which metric to fill the matrix with
\itemize{
 \item{\code{"ADJ_ZSTAT"} : }{MAGMA Z-statistic after adjusting for
 gene length and other confounds (\emph{DEFAULT}).}
 \item{\code{"ZSTAT"} : }{Unadjusted MAGMA Z-statistic.}
 \item{\code{"P"} : }{MAGMA P-value.}
}}

\item{drop_mhc}{Drop genes from the MHC
(Major Histocompatibility Complex) region.}

\item{save_path}{Path to save results as RDS object.}

\item{fillna}{Value to fill \code{NA}s with.}

\item{trans_fun}{Transformation function.}

\item{agg_fun}{Aggregation function.}

\item{all}{logical; \code{all = TRUE} is shorthand to save setting both
\code{all.x = TRUE} and \code{all.y = TRUE}.}

\item{as_matrix}{Convert the data to a sparse matrix.}

\item{verbose}{Print messages.}

\item{nCores}{Number of cores to parallelise across.}
}
\value{
Sparse matrix.
}
\description{
Once you've run \link[MAGMA.Celltyping]{map_snps_to_genes} 
on multiple GWAS/QTL,
you can aggregate the gene-wise results into a single gene x trait matrix.
}
\examples{
meta <- opengwas_meta()
meta <- meta[grepl("eqtl-a-",id),][seq(5)]
X <- magma_matrix(magma_out = meta$magma_out, workers=10) 
}
