% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomix-package.R
\docType{package}
\name{phenomix-package}
\alias{phenomix}
\alias{phenomix-package}
\title{phenomix: Fit Density Curves to Peak Timing Data that Varies over Time}
\description{
The 'phenomix' package fits time-varying density curves to run timing type data commonly encountered in fisheries and ecology. Example applications include to peak run timing curves collected for juvenile or adult Pacific salmon, though could also be applied to other kinds of data such as hydrographs, plant phenology (flowering, leaf out).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nwfsc-cb.github.io/phenomix}
  \item \url{https://github.com/nwfsc-cb/phenomix}
}

}
\author{
\strong{Maintainer}: Eric J. Ward \email{eric.ward@noaa.gov}

Other contributors:
\itemize{
  \item Samantha M. Wilson [contributor]
  \item Joseph H. Anderson [contributor]
}

}
\keyword{internal}
