% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag_enrich.R
\name{plot_dag_enrich}
\alias{plot_dag_enrich}
\title{Plot enrichment of terms in a DAG ontology}
\usage{
plot_dag_enrich(
  obj,
  ont,
  min_hits = 3,
  min_offspring = 100,
  min_depth = NULL,
  exact_depth = NULL,
  exact_ont_ancestors = FALSE,
  cluster_col = "seurat_clusters",
  label_col = "dag_enrich.name_wrap",
  color_col = cluster_col,
  replace_char = list(. = ":", `_` = ":"),
  q_threshold = 0.05,
  preferred_palettes = "kovesi.cyclic_mygbm_30_95_c78",
  add_labels = TRUE,
  reduction = NULL,
  label_alpha = 0.5,
  point_alpha = 0.4,
  point_size = 0.7,
  label_size = 3,
  label_width = 20,
  repel = TRUE,
  box = TRUE,
  box.padding = 1,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{repel}{Use \code{geom_text_repel} to create nicely-repelled labels}

\item{box}{Use geom_label/geom_label_repel (includes a box around the text
labels)}

\item{...}{
  Arguments passed on to \code{\link[Seurat:LabelClusters]{Seurat::LabelClusters}}
  \describe{
    \item{\code{plot}}{A ggplot2-based scatter plot}
    \item{\code{id}}{Name of variable used for coloring scatter plot}
    \item{\code{clusters}}{Vector of cluster ids to label}
    \item{\code{labels}}{Custom labels for the clusters}
    \item{\code{split.by}}{Split labels by some grouping label, useful when using
\code{\link[ggplot2]{facet_wrap}} or \code{\link[ggplot2]{facet_grid}}}
    \item{\code{geom}}{Name of geom to get X/Y aesthetic names for}
    \item{\code{position}}{How to place the label if repel = FALSE. If "median", place
the label at the median position. If "nearest" place the label at the
position of the nearest data point to the median.}
  }}
}
\description{
Plot enrichment of terms in a DAG ontology.
}
\examples{
ont <- KGExplorer::get_ontology("hp")
obj <- get_HPO()
out <- plot_dag_enrich(obj, ont)
}
