% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plotting function to be called by user}
\usage{
plot_diagnostics(fitted, type = "timing")
}
\arguments{
\item{fitted}{A fitted model object}

\item{type}{A plot type for ggplot, either "timing" or "scatter"}
}
\description{
These functions make some basic plots for the user
}
