% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_bar.R
\name{plot_enrichment_bar}
\alias{plot_enrichment_bar}
\title{Plot enrichment results: barplot}
\usage{
plot_enrichment_bar(
  res,
  qvalue_thresh = 0.05,
  show_plot = TRUE,
  save_path = NULL,
  verbose = TRUE
)
}
\description{
Plot enrichment results: barplot
}
\keyword{internal}
