% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ontological_velocity.R
\name{plot_ontological_velocity}
\alias{plot_ontological_velocity}
\title{Plot ontological velocity}
\usage{
plot_ontological_velocity(
  obj,
  ont = KGExplorer::get_ontology("hp"),
  id_col = NULL,
  iterations = 5000,
  k = 30,
  agg_fun = stats::weighted.mean,
  show_plot = TRUE
)
}
\description{
Plot ontological velocity.
}
\examples{
obj <- get_HPO()
ont <- KGExplorer::get_ontology("hp")
out <- plot_ontological_velocity(obj,ont)
}
