% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{Plot reduction}
\usage{
plot_reduction(
  obj,
  keys = NULL,
  metadata = NULL,
  x_dim = 1,
  y_dim = 2,
  fix_rownames = FALSE,
  color_var = NULL,
  label_var = NULL,
  size_var = NULL,
  labels = !is.null(label_var),
  max.overlaps = 30,
  show_plot = TRUE,
  point_alpha = 0.6
)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{AnnData}}
\item{\code{prcomp}}
\item{\code{CellDataSet}}
\item{\code{list}}
\item{\code{matrix/data.frame}}
}}

\item{keys}{The keys of reductions to extract from.}

\item{metadata}{Metadata associated with \code{obj}. 
Will be automatically extracted from \code{obj} if it is a
 \code{Seurat} object.}

\item{fix_rownames}{Replace non-ASCII characters in row names.}

\item{color_var}{\code{metadata} variable to use as point color.}

\item{label_var}{\code{metadata} variable to use as point labels}

\item{size_var}{\code{metadata} variable to use as point size.}

\item{labels}{Whether to show labels using \link[ggrepel]{geom_text_repel}.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
Defaults to 10.}

\item{show_plot}{Whether to print the plot.}

\item{point_alpha}{Opacity of data points.}
}
\value{
ggplot object
}
\description{
Plot dimensionality reduction results
from \code{run_<method>} or a Seurat object.
}
