% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qthill}
\alias{qthill}
\title{Calculate the quantile of a Student-t distribution}
\usage{
qthill(p, mean, sd, df)
}
\arguments{
\item{p}{The quantile of the distribution}

\item{mean}{The mean of the distribution}

\item{sd}{The std deviation of the distribution}

\item{df}{The df of the distribution}
}
\description{
Calculate the quantile of a Student-t distribution
}
