% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_matrix.R
\name{reconstruct_matrix}
\alias{reconstruct_matrix}
\title{Reconstruct original matrix from DeGAs components}
\usage{
reconstruct_matrix(
  u_name = "contribution_phe",
  v_name = "contribution_var",
  d_name = NULL,
  u_labels = "label_phe_code",
  v_labels = if (gene_level) {     "label_gene" } else {     "label_var" },
  npz_obj = NULL,
  annot = NULL,
  gene_level = FALSE,
  remove_nongenes = TRUE,
  filter_quantiles = FALSE,
  norm = TRUE,
  norm_scale = c(1, 100),
  replace_underscores = c(u = TRUE, v = TRUE),
  translate_snp_positions = FALSE,
  invert_UV_labels = FALSE,
  verbose = TRUE
)
}
\description{
Reconstruct original matrix from DeGAs components
}
\keyword{internal}
