% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_matrix.R
\name{reconstruct_matrix}
\alias{reconstruct_matrix}
\title{Reconstruct original matrix from DeGAs components}
\source{

}
\usage{
reconstruct_matrix(
  u_name = "contribution_phe",
  v_name = "contribution_var",
  d_name = NULL,
  DEGAS = py$DEGAS,
  annot = NULL,
  gene_level = F,
  remove_nongenes = T,
  filter_quantiles = F,
  norm = T,
  norm_scale = c(1, 100),
  replace_underscores = T
)
}
\description{
Reconstruct original matrix from DeGAs components
}
