% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mofa2.R
\name{run_mofa2}
\alias{run_mofa2}
\title{Run MOFA2}
\source{
\href{https://biofam.github.io/MOFA2/}{MOFA2 site}
}
\usage{
run_mofa2(
  obj,
  groups = NULL,
  metadata = NULL,
  assay = NULL,
  slot = NULL,
  features = NULL,
  metadata_idcol = "sample",
  transpose = FALSE,
  maxiter = 1000,
  data_options = NULL,
  model_options = NULL,
  training_options = NULL,
  stochastic_options = NULL,
  mefisto_options = NULL,
  outfile = file.path(tempdir(), "model.hdf5"),
  reductions = c("umap"),
  seed = 2020,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{Named list of input data matrices, or a \pkg{Seurat} object.}

\item{groups}{group information, only relevant when using the multi-group framework.}

\item{metadata}{Metadata associated with \code{obj}. 
Will be extracted automatically if \code{obj} is a \pkg{Seurat} object.}

\item{assay}{Assay to use in \code{obj}.}

\item{slot}{Slot to use in \code{obj}.}

\item{features}{Features to use in reduction. 
If \code{NULL} and \code{obj} is a \pkg{Seurat} object, 
will use precalculated variable genes.}

\item{metadata_idcol}{The name of the column to add to \code{metadata} 
after fixing IDs.}

\item{transpose}{First transpose each matrix in \code{mat_list}.}

\item{maxiter}{Maximum number of training iterations (\emph{DEFAULT} = 1000).}

\item{data_options}{list of data_options (see \code{\link[MOFA2]{get_default_data_options}} details). 
If NULL, default options are used.}

\item{model_options}{list of model options (see \code{\link[MOFA2]{get_default_model_options}} for details). 
If NULL, default options are used.}

\item{training_options}{list of training options (see \code{\link[MOFA2]{get_default_training_options}} for details). 
If NULL, default options are used.}

\item{stochastic_options}{list of options for stochastic variational inference (see \code{\link[MOFA2]{get_default_stochastic_options}} for details). 
If NULL, default options are used.}

\item{mefisto_options}{list of options for mefisto (see \code{\link[MOFA2]{get_default_mefisto_options}} for details). 
If NULL, default options are used.}

\item{outfile}{output file for the model (.hdf5 format). If \code{NULL}, a temporary file is created.}

\item{reductions}{In addition to returning  MOFA factors,
further reduce the data with "umap" and/or "tsne".}

\item{seed}{Seed passed to \link[base]{set.seed}
for reproducibility between runs.}

\item{verbose}{Print messages.}

\item{...}{Additional parameters passed to \link[MOFA2]{run_mofa}.}
}
\description{
Run Multi-Omics Factor Analysis v2 (MOFA2).
}
\details{
Uses \link[MOFA2]{run_mofa} to reduce a single dataset,
or multiple views of the same dataset, into a
set number of factors.
}
