% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_variancePartition.R
\name{run_variancePartition}
\alias{run_variancePartition}
\title{Run \pkg{variancePartition}}
\usage{
run_variancePartition(
  obj,
  metadata = NULL,
  nCores = NULL,
  form = NULL,
  is_opengwas = FALSE,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{Matrix or \pkg{Seurat} object to run
\link[variancePartition]{fitExtractVarPartModel} on.}

\item{metadata}{\link[data.frame]{data.frame} containing metadata
to use in \code{form}. If \code{obj} is a \pkg{Seurat} object,
this can be left \code{NULL} and metadata 
will be extracted automatically.}

\item{nCores}{Number of cores to use in parallel.}

\item{is_opengwas}{Whether the data comes from
 \href{https://gwas.mrcieu.ac.uk/}{OpenGWAS}. 
If \code{TRUE}, a predefined formula and metadata 
processing procedure will be used.}

\item{...}{Additional arguments passed to 
\link[variancePartition]{fitExtractVarPartModel}.}
}
\description{
Run \link[variancePartition]{fitExtractVarPartModel} 
to evaluate the effect of each metadata variable on your data.
}
